refactor: split MySQL CLI runtime
