+++ src/infra/adapters/mysql.rs
@@ -1,73 +1,36 @@
-use std::ffi::OsStr;
-#[cfg(test)]
-use std::fs;
-use std::io::{self, Write};
 use std::path::PathBuf;
-use std::pin::Pin;
-use std::process::Stdio;
-use std::sync::Arc;
-use std::task::{Context, Poll};
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
-#[cfg(unix)]
-use std::os::fd::{AsRawFd, FromRawFd};
-
-use async_trait::async_trait;
-use quick_xml::Reader;
-use quick_xml::escape::unescape;
-use quick_xml::events::{BytesRef, Event};
-use serde::Deserialize;
-#[cfg(unix)]
-use tokio::fs::File as TokioFile;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncWriteExt, BufReader, ReadBuf};
-use tokio::process::Child;
-use tokio::process::Command;
-#[cfg(not(unix))]
-use tokio::process::{ChildStderr, ChildStdin, ChildStdout};
-use tokio::time::timeout;
-use uuid::Uuid;
-
-#[cfg(any(test, feature = "test-support"))]
+use crate::adapters::csv_export::export_to_downloads;
+#[cfg(feature = "test-support")]
 use crate::adapters::csv_export::export_to_path;
-use crate::adapters::csv_export::{CsvFileWriter, export_to_downloads};
-#[cfg(test)]
-use crate::app::policy::sql::mysql_statement::split_mysql_statements;
-use crate::app::policy::sql::mysql_statement::{
-    MysqlStatement, MysqlStatementKind, classify_mysql_statement, has_mysql_read_only_side_effect,
-    mysql_tree_explain_query_kind,
-};
-use crate::app::policy::write::sql_risk::{
-    MultiStatementDecision, evaluate_mysql_multi_statement, mysql_statement_is_data_modifying,
-    mysql_statement_is_schema_modifying,
-};
+use crate::app::policy::sql::mysql_statement::mysql_tree_explain_query_kind;
 use crate::app::ports::outbound::{
-    AccessMode, ConnectionProbe, DatabaseCli, DbOperationError, DdlGenerator, DsnBuilder,
-    MYSQL_CLI_VERSION_REQUIRED_MARKER, MYSQL_SERVER_VERSION_REQUIRED_MARKER,
-    MYSQL_SQL_MODE_UNSUPPORTED_MARKER, QueryExecutor, SqlDialect,
+    AccessMode, ConnectionProbe, DbOperationError, DdlGenerator, DsnBuilder,
+    MYSQL_CLI_VERSION_REQUIRED_MARKER, QueryExecutor, SqlDialect,
 };
 use crate::domain::connection::{ConnectionProfile, DatabaseType};
-use crate::domain::{
-    CommandTag, QueryResult, QuerySource, QueryValue, RefreshScope, Table, WriteExecutionResult,
-};
+use crate::domain::{QueryResult, QuerySource, QueryValue, Table, WriteExecutionResult};
+use async_trait::async_trait;
 
+mod cli;
 mod dsn;
 mod metadata;
 mod option_file;
 mod sql;
 
-use dsn::{
-    MySqlDsn, build_mysql_dsn, parse_mysql_dsn, validate_mysql_tls_files, validate_mysql_values,
+use cli::{
+    MYSQL_QUERY_TIMEOUT, MySqlProbeResponse, MysqlMetadataSession, MysqlResultSet,
+    classify_mysql_probe_failure, clean_stderr, export_mysql_csv_to_file,
+    is_oracle_mysql_cli_84_version, mysql_probe_args, run_mysql_adhoc, run_mysql_command,
+    run_mysql_single_statement, validate_mysql_export_query, validate_mysql_multi_query,
+    validate_server_version, validate_sql_mode,
 };
+
+use dsn::{build_mysql_dsn, parse_mysql_dsn, validate_mysql_tls_files, validate_mysql_values};
 use option_file::MySqlOptionFile;
 
 pub struct MySqlAdapter;
-
-const MYSQL_PROBE_TIMEOUT: Duration = Duration::from_secs(11);
-const MYSQL_QUERY_TIMEOUT: Duration = Duration::from_secs(31);
-const MYSQL_EXPORT_TIMEOUT: Duration = Duration::from_secs(MYSQL_QUERY_TIMEOUT.as_secs() * 10);
-const MYSQL_PROBE_QUERY: &str = "SELECT JSON_OBJECT('database', DATABASE(), 'user', CURRENT_USER(), 'version', VERSION(), 'sql_mode', @@SESSION.sql_mode)";
-const MYSQL_READ_ONLY_STATEMENT: &str = "SET SESSION TRANSACTION READ ONLY";
-const MYSQL_SESSION_MARKER_COLUMN: &str = "__sabiql_session_marker";
 
 impl MySqlAdapter {
     pub fn new() -> Self {
@@ -87,25 +50,7 @@ pub async fn run_mysql_cli_script_for_test(
     dsn: &str,
     script: &str,
 ) -> Result<Vec<u8>, DbOperationError> {
-    let target = parse_mysql_dsn(dsn)?;
-    validate_mysql_values(&target)?;
-    validate_mysql_tls_files(&target)?;
-    let option_file = MySqlOptionFile::create(&target)?;
-    let mut process = MysqlProcess::spawn_with_program(OsStr::new("mysql"), &option_file.path)?;
-    let result = async {
-        write_mysql_input(&mut process, script.as_bytes()).await?;
-        write_mysql_input(&mut process, b"\x04").await?;
-        read_pty_all(&mut process.pty)
-            .await
-            .map_err(|error| DbOperationError::QueryFailed(error.to_string()))
-    }
-    .await;
-    if result.is_err() {
-        cleanup_mysql_process(&mut process).await;
-    } else {
-        let _ = process.child.wait().await;
-    }
-    result
+    cli::run_mysql_cli_script_for_test(dsn, script).await
 }
 
 #[cfg(feature = "test-support")]
@@ -487,4235 +432,5 @@ impl MySqlAdapter {
     ) -> Result<std::process::Output, DbOperationError> {
         let args = mysql_probe_args(option_file);
         run_mysql_command(args, Some(option_file)).await
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct MySqlProbeResponse {
-    database: Option<String>,
-    user: String,
-    version: String,
-    sql_mode: String,
-}
-
-fn contains_unsupported_mysql_product(value: &str) -> bool {
-    let lower = value.to_ascii_lowercase();
-    ["mariadb", "percona", "tidb", "vitess", "aurora"]
-        .iter()
-        .any(|product| lower.contains(product))
-}
-
-fn is_oracle_mysql_cli_84_version(value: &str) -> bool {
-    let lower = value.to_ascii_lowercase();
-    lower.contains("mysql")
-        && !contains_unsupported_mysql_product(value)
-        && version_major_minor(value) == Some((8, 4))
-}
-
-fn is_oracle_mysql_server_84_version(value: &str) -> bool {
-    !contains_unsupported_mysql_product(value) && version_major_minor(value) == Some((8, 4))
-}
-
-fn validate_server_version(version: &str) -> Result<(), DbOperationError> {
-    if is_oracle_mysql_server_84_version(version) {
-        Ok(())
-    } else {
-        Err(DbOperationError::UnsupportedOperation(format!(
-            "{MYSQL_SERVER_VERSION_REQUIRED_MARKER}: {version}"
-        )))
-    }
-}
-
-fn validate_sql_mode(sql_mode: &str) -> Result<(), DbOperationError> {
-    let unsupported = sql_mode.split(',').map(str::trim).any(|mode| {
-        mode.eq_ignore_ascii_case("NO_BACKSLASH_ESCAPES")
-            || mode.eq_ignore_ascii_case("ANSI_QUOTES")
-    });
-    if unsupported {
-        Err(DbOperationError::UnsupportedOperation(format!(
-            "{MYSQL_SQL_MODE_UNSUPPORTED_MARKER}: {sql_mode}"
-        )))
-    } else {
-        Ok(())
-    }
-}
-
-fn version_major_minor(value: &str) -> Option<(u32, u32)> {
-    let mut numbers = value
-        .split(|character: char| !character.is_ascii_digit())
-        .filter(|part| !part.is_empty())
-        .filter_map(|part| part.parse::<u32>().ok());
-    Some((numbers.next()?, numbers.next()?))
-}
-
-fn mysql_probe_args(option_file: &std::path::Path) -> Vec<String> {
-    vec![
-        format!("--defaults-file={}", option_file.display()),
-        "--no-login-paths".to_string(),
-        "--protocol=TCP".to_string(),
-        "--connect-timeout=10".to_string(),
-        "--batch".to_string(),
-        "--raw".to_string(),
-        "--skip-column-names".to_string(),
-        "--binary-mode".to_string(),
-        "--skip-reconnect".to_string(),
-        format!("--execute={MYSQL_PROBE_QUERY}"),
-    ]
-}
-
-async fn run_mysql_command<I, S>(
-    args: I,
-    option_file: Option<&PathBuf>,
-) -> Result<std::process::Output, DbOperationError>
-where
-    I: IntoIterator<Item = S>,
-    S: AsRef<OsStr>,
-{
-    let mut command = Command::new("mysql");
-    command
-        .args(args)
-        .stdin(Stdio::null())
-        .env_remove("MYSQL_PWD")
-        .env_remove("MYSQL_PASSWORD")
-        .kill_on_drop(true);
-    if option_file.is_some() {
-        command.stdout(Stdio::piped()).stderr(Stdio::piped());
-    }
-
-    match timeout(MYSQL_PROBE_TIMEOUT, command.output()).await {
-        Ok(Ok(output)) => Ok(output),
-        Ok(Err(error)) if error.kind() == io::ErrorKind::NotFound => {
-            Err(DbOperationError::CommandNotFound {
-                command: DatabaseCli::MySql,
-                details: error.to_string(),
-            })
-        }
-        Ok(Err(error)) => Err(DbOperationError::ConnectionFailed(error.to_string())),
-        Err(_) => Err(DbOperationError::Timeout(
-            "mysql probe exceeded the connection timeout".to_string(),
-        )),
-    }
-}
-
-fn clean_stderr(stderr: &[u8]) -> String {
-    let text = String::from_utf8_lossy(stderr).trim().to_string();
-    if text.is_empty() {
-        "mysql probe failed".to_string()
-    } else {
-        text
-    }
-}
-
-fn classify_mysql_probe_failure(stderr: String) -> DbOperationError {
-    if is_mysql_connect_timeout_message(&stderr) {
-        DbOperationError::Timeout(stderr)
-    } else {
-        DbOperationError::ConnectionFailed(stderr)
-    }
-}
-
-fn is_mysql_connect_timeout_message(value: &str) -> bool {
-    let lower = value.to_ascii_lowercase();
-    lower.contains("can't connect to mysql server")
-        && (lower.contains("(110)") || lower.contains("(10060)"))
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct MysqlResultSet {
-    columns: Vec<String>,
-    values: Vec<Vec<QueryValue>>,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum MysqlMetadataFallbackKind {
-    Select,
-    Show,
-    Describe,
-}
-
-fn mysql_metadata_fallback_kind(kind: &MysqlStatementKind) -> Option<MysqlMetadataFallbackKind> {
-    match kind {
-        MysqlStatementKind::Select => Some(MysqlMetadataFallbackKind::Select),
-        MysqlStatementKind::Show => Some(MysqlMetadataFallbackKind::Show),
-        MysqlStatementKind::Describe => Some(MysqlMetadataFallbackKind::Describe),
-        _ => None,
-    }
-}
-
-fn mysql_metadata_select_query(
-    query: &str,
-    source_alias: &str,
-    marker_alias: &str,
-) -> Result<String, DbOperationError> {
-    let query = query.trim().trim_end_matches(';').trim_end();
-    if query.is_empty() {
-        return Err(DbOperationError::QueryFailed(
-            "MySQL empty SELECT cannot be used for metadata fallback".to_string(),
-        ));
-    }
-    if mysql_metadata_select_has_unproven_function(query) {
-        return Err(DbOperationError::UnsupportedOperation(
-            "MySQL SELECT metadata fallback cannot prove that function calls are side-effect free"
-                .to_string(),
-        ));
-    }
-    if has_mysql_read_only_side_effect(query)
-        .map_err(|error| DbOperationError::QueryFailed(error.to_string()))?
-    {
-        return Err(DbOperationError::UnsupportedOperation(
-            "MySQL SELECT metadata fallback cannot prove that the query is side-effect free"
-                .to_string(),
-        ));
-    }
-    Ok(sql::build_metadata_select_query(
-        query,
-        source_alias,
-        marker_alias,
-    ))
-}
-
-fn mysql_metadata_select_has_unproven_function(sql: &str) -> bool {
-    let bytes = sql.as_bytes();
-    let mut index = 0;
-    while index < bytes.len() {
-        if bytes[index] == b'#'
-            || (bytes.get(index..index + 2) == Some(b"--")
-                && bytes.get(index + 2).is_none_or(u8::is_ascii_whitespace))
-        {
-            index = bytes[index..]
-                .iter()
-                .position(|byte| *byte == b'\n')
-                .map_or(bytes.len(), |offset| index + offset + 1);
-            continue;
-        }
-        if bytes.get(index..index + 2) == Some(b"/*") {
-            index = bytes
-                .get(index + 2..)
-                .and_then(|rest| rest.windows(2).position(|window| window == b"*/"))
-                .map_or(bytes.len(), |offset| index + offset + 4);
-            continue;
-        }
-        if bytes[index] == b'@' {
-            return true;
-        }
-        if matches!(bytes[index], b'\'' | b'"' | b'`') {
-            let quote = bytes[index];
-            index = skip_mysql_metadata_quoted(bytes, index, quote);
-            let next = skip_mysql_metadata_trivia(bytes, index);
-            if quote == b'`' && bytes.get(next) == Some(&b'(') {
-                return true;
-            }
-            continue;
-        }
-        if bytes[index].is_ascii_alphabetic() || matches!(bytes[index], b'_' | b'$') {
-            let start = index;
-            index += 1;
-            while index < bytes.len()
-                && (bytes[index].is_ascii_alphanumeric() || matches!(bytes[index], b'_' | b'$'))
-            {
-                index += 1;
-            }
-            let next = skip_mysql_metadata_trivia(bytes, index);
-            if bytes.get(next) == Some(&b'(') {
-                let name = &sql[start..index];
-                let cte_column_list = mysql_metadata_is_cte_column_list(sql, next);
-                let qualified = mysql_metadata_has_qualifier(bytes, start);
-                if !cte_column_list
-                    && (qualified
-                        || !name.eq_ignore_ascii_case("SLEEP")
-                            && !matches!(
-                                name.to_ascii_uppercase().as_str(),
-                                "AS" | "CASE" | "IN" | "EXISTS" | "OVER"
-                            ))
-                {
-                    return true;
-                }
-            }
-            continue;
-        }
-        index += 1;
-    }
-    false
-}
-
-fn skip_mysql_metadata_trivia(bytes: &[u8], mut index: usize) -> usize {
-    loop {
-        while bytes.get(index).is_some_and(u8::is_ascii_whitespace) {
-            index += 1;
-        }
-        if bytes.get(index) == Some(&b'#')
-            || (bytes.get(index..index + 2) == Some(b"--")
-                && bytes.get(index + 2).is_some_and(u8::is_ascii_whitespace))
-        {
-            index = bytes[index..]
-                .iter()
-                .position(|byte| *byte == b'\n')
-                .map_or(bytes.len(), |offset| index + offset + 1);
-            continue;
-        }
-        if bytes.get(index..index + 2) == Some(b"/*") {
-            index = bytes
-                .get(index + 2..)
-                .and_then(|rest| rest.windows(2).position(|window| window == b"*/"))
-                .map_or(bytes.len(), |offset| index + offset + 4);
-            continue;
-        }
-        return index;
-    }
-}
-
-fn mysql_metadata_has_qualifier(bytes: &[u8], end: usize) -> bool {
-    let mut index = 0;
-    let mut previous = None;
-    while index < end {
-        if bytes[index] == b'#'
-            || (bytes.get(index..index + 2) == Some(b"--")
-                && bytes.get(index + 2).is_some_and(u8::is_ascii_whitespace))
-            || bytes.get(index..index + 2) == Some(b"/*")
-        {
-            let next = skip_mysql_metadata_trivia(bytes, index);
-            if next == index {
-                index += 1;
-            } else {
-                index = next;
-            }
-            continue;
-        }
-        if matches!(bytes[index], b'\'' | b'"' | b'`') {
-            index = skip_mysql_metadata_quoted(bytes, index, bytes[index]);
-            previous = Some(b'\'');
-        } else if bytes[index].is_ascii_whitespace() {
-            index += 1;
-        } else {
-            previous = Some(bytes[index]);
-            index += 1;
-        }
-    }
-    previous == Some(b'.')
-}
-
-fn mysql_metadata_is_cte_column_list(sql: &str, candidate_open: usize) -> bool {
-    let bytes = sql.as_bytes();
-    let mut index = skip_mysql_metadata_trivia(bytes, 0);
-    let Some(after_with) = mysql_metadata_keyword_end(bytes, index, "WITH") else {
-        return false;
-    };
-    index = skip_mysql_metadata_trivia(bytes, after_with);
-    if let Some(after_recursive) = mysql_metadata_keyword_end(bytes, index, "RECURSIVE") {
-        index = skip_mysql_metadata_trivia(bytes, after_recursive);
-    }
-
-    loop {
-        index = match mysql_metadata_cte_name_end(bytes, index) {
-            Some(end) => skip_mysql_metadata_trivia(bytes, end),
-            None => return false,
-        };
-        let mut column_list = None;
-        if bytes.get(index) == Some(&b'(') {
-            let Some(end) = mysql_metadata_parenthesized_end(bytes, index) else {
-                return false;
-            };
-            column_list = Some(index);
-            index = skip_mysql_metadata_trivia(bytes, end);
-        }
-        let Some(after_as) = mysql_metadata_keyword_end(bytes, index, "AS") else {
-            return false;
-        };
-        index = skip_mysql_metadata_trivia(bytes, after_as);
-        let Some(body_end) = bytes
-            .get(index)
-            .filter(|byte| **byte == b'(')
-            .and_then(|_| mysql_metadata_parenthesized_end(bytes, index))
-        else {
-            return false;
-        };
-        if column_list == Some(candidate_open) {
-            return true;
-        }
-        index = skip_mysql_metadata_trivia(bytes, body_end);
-        if bytes.get(index) != Some(&b',') {
-            return false;
-        }
-        index = skip_mysql_metadata_trivia(bytes, index + 1);
-    }
-}
-
-fn mysql_metadata_cte_name_end(bytes: &[u8], index: usize) -> Option<usize> {
-    if bytes.get(index) == Some(&b'`') {
-        Some(skip_mysql_metadata_quoted(bytes, index, b'`'))
-    } else if bytes
-        .get(index)
-        .is_some_and(|byte| byte.is_ascii_alphabetic() || matches!(byte, b'_' | b'$'))
-    {
-        Some(
-            index
-                + 1
-                + bytes[index + 1..]
-                    .iter()
-                    .take_while(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'$'))
-                    .count(),
-        )
-    } else {
-        None
-    }
-}
-
-fn mysql_metadata_keyword_end(bytes: &[u8], index: usize, keyword: &str) -> Option<usize> {
-    let end = index.checked_add(keyword.len())?;
-    if bytes
-        .get(index..end)?
-        .eq_ignore_ascii_case(keyword.as_bytes())
-        && !bytes
-            .get(end)
-            .is_some_and(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'$'))
-    {
-        Some(end)
-    } else {
-        None
-    }
-}
-
-fn mysql_metadata_parenthesized_end(bytes: &[u8], open: usize) -> Option<usize> {
-    let mut depth = 0;
-    let mut index = open;
-    while index < bytes.len() {
-        if matches!(bytes[index], b'\'' | b'"' | b'`') {
-            index = skip_mysql_metadata_quoted(bytes, index, bytes[index]);
-            continue;
-        }
-        let next = skip_mysql_metadata_trivia(bytes, index);
-        if next != index {
-            index = next;
-            continue;
-        }
-        match bytes[index] {
-            b'(' => depth += 1,
-            b')' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(index + 1);
-                }
-            }
-            _ => {}
-        }
-        index += 1;
-    }
-    None
-}
-
-fn skip_mysql_metadata_quoted(bytes: &[u8], start: usize, quote: u8) -> usize {
-    let mut index = start + 1;
-    while index < bytes.len() {
-        if bytes[index] == b'\\' && quote != b'`' {
-            index = (index + 2).min(bytes.len());
-        } else if bytes[index] == quote {
-            if bytes.get(index + 1) == Some(&quote) {
-                index += 2;
-            } else {
-                return index + 1;
-            }
-        } else {
-            index += 1;
-        }
-    }
-    bytes.len()
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct MysqlExecutionResult {
-    result_set: Option<MysqlResultSet>,
-    command_tag: Option<CommandTag>,
-    refresh_scope: RefreshScope,
-}
-
-struct MysqlCommandEvent {
-    kind: MysqlStatementKind,
-    target: Option<String>,
-    tag: CommandTag,
-}
-
-fn validate_mysql_multi_query(
-    query: &str,
-    selected_database: Option<&str>,
-    access_mode: AccessMode,
-) -> Result<Vec<MysqlStatement>, DbOperationError> {
-    let decision = evaluate_mysql_multi_statement(query, selected_database);
-    let (statements, risk) = match decision {
-        MultiStatementDecision::Allow { statements, risk } => (statements, risk),
-        MultiStatementDecision::Block { reason } => {
-            return Err(DbOperationError::UnsupportedOperation(reason));
-        }
-    };
-    if access_mode.is_read_only() && !risk.read_only_allowed {
-        return Err(DbOperationError::PermissionDenied(
-            "read-only mode blocks MySQL write statements".to_string(),
-        ));
-    }
-    statements
-        .iter()
-        .map(|statement| {
-            classify_mysql_statement(statement)
-                .map_err(|error| DbOperationError::UnsupportedOperation(error.to_string()))
-        })
-        .collect()
-}
-
-fn mysql_metadata_fallback_has_unsupported_session_state(statements: &[MysqlStatement]) -> bool {
-    let mut temporary_table_created = false;
-    for statement in statements {
-        if temporary_table_created
-            && matches!(
-                statement.kind,
-                MysqlStatementKind::Show | MysqlStatementKind::Describe
-            )
-        {
-            return true;
-        }
-        if matches!(
-            statement.kind,
-            MysqlStatementKind::CreateTable { temporary: true }
-        ) {
-            temporary_table_created = true;
-        }
-    }
-    false
-}
-
-fn validate_mysql_export_query(
-    query: &str,
-    selected_database: Option<&str>,
-) -> Result<(), DbOperationError> {
-    let statements = validate_mysql_multi_query(query, selected_database, AccessMode::ReadOnly)?;
-    if statements.len() != 1
-        || !matches!(
-            statements[0].kind,
-            MysqlStatementKind::Select
-                | MysqlStatementKind::Table
-                | MysqlStatementKind::Show
-                | MysqlStatementKind::Describe
-        )
-    {
-        return Err(DbOperationError::UnsupportedOperation(
-            "MySQL CSV export supports a single read-only result query".to_string(),
-        ));
-    }
-    Ok(())
-}
-
-struct MysqlProcess {
-    child: Child,
-    #[cfg(unix)]
-    pty: MysqlPty,
-    #[cfg(not(unix))]
-    stdin: ChildStdin,
-    #[cfg(not(unix))]
-    stdout: ChildStdout,
-    #[cfg(not(unix))]
-    stderr: ChildStderr,
-    #[cfg(not(unix))]
-    pending: Vec<u8>,
-    #[cfg(not(unix))]
-    pending_stderr: Vec<u8>,
-    #[cfg(not(unix))]
-    frame_scanner: MysqlResultsetFrameScanner,
-}
-
-const MYSQL_RESULTSET_START: &[u8] = b"<resultset";
-const MYSQL_RESULTSET_END: &[u8] = b"</resultset>";
-
-#[derive(Debug, Default)]
-struct MysqlResultsetFrameScanner {
-    resultset_start: Option<usize>,
-    resultset_end: Option<usize>,
-    resultset_start_cursor: usize,
-    resultset_end_cursor: usize,
-}
-
-impl MysqlResultsetFrameScanner {
-    fn frame_bounds(&mut self, buffer: &[u8]) -> Option<(usize, usize)> {
-        if self.resultset_start_cursor > buffer.len() {
-            self.resultset_start_cursor = 0;
-        }
-        if self.resultset_end_cursor > buffer.len() {
-            self.resultset_end_cursor = 0;
-        }
-
-        if self.resultset_start.is_none() {
-            let scan_start = self
-                .resultset_start_cursor
-                .saturating_sub(MYSQL_RESULTSET_START.len().saturating_sub(1));
-            self.resultset_start = find_bytes_from(buffer, MYSQL_RESULTSET_START, scan_start);
-            if let Some(start) = self.resultset_start {
-                self.resultset_end_cursor = start;
-            } else {
-                self.resultset_start_cursor = buffer.len();
-                return None;
-            }
-        }
-
-        let start = self.resultset_start?;
-        if self.resultset_end.is_none() {
-            let scan_start = self
-                .resultset_end_cursor
-                .saturating_sub(MYSQL_RESULTSET_END.len().saturating_sub(1))
-                .max(start);
-            self.resultset_end = find_bytes_from(buffer, MYSQL_RESULTSET_END, scan_start)
-                .map(|end| end + MYSQL_RESULTSET_END.len());
-            if self.resultset_end.is_none() {
-                self.resultset_end_cursor = buffer.len();
-                return None;
-            }
-        }
-
-        self.resultset_end.map(|end| (start, end))
-    }
-
-    #[cfg(any(not(unix), test))]
-    fn take(&mut self, buffer: &mut Vec<u8>) -> Option<Vec<u8>> {
-        let bounds = self.frame_bounds(buffer)?;
-        Some(self.take_bounds(buffer, bounds))
-    }
-
-    fn take_bounds(&mut self, buffer: &mut Vec<u8>, (start, end): (usize, usize)) -> Vec<u8> {
-        let frame = buffer[start..end].to_vec();
-        buffer.drain(..end);
-        self.resultset_start_cursor = self
-            .resultset_start_cursor
-            .saturating_sub(end)
-            .min(buffer.len());
-        self.resultset_end_cursor = self
-            .resultset_end_cursor
-            .saturating_sub(end)
-            .min(buffer.len());
-        self.resultset_start = None;
-        self.resultset_end = None;
-        frame
-    }
-
-    #[cfg(unix)]
-    fn reset(&mut self) {
-        *self = Self::default();
-    }
-}
-
-#[cfg(unix)]
-struct MysqlPty {
-    input: TokioFile,
-    output: TokioFile,
-    pending: Vec<u8>,
-    frame_scanner: MysqlResultsetFrameScanner,
-}
-
-impl MysqlProcess {
-    fn spawn_with_program(
-        program: &OsStr,
-        option_file: &std::path::Path,
-    ) -> Result<Self, DbOperationError> {
-        #[cfg(unix)]
-        {
-            Self::spawn_with_pty(program, option_file)
-        }
-
-        #[cfg(not(unix))]
-        {
-            let mut command = Command::new(program);
-            command
-                .args(mysql_query_args(option_file))
-                .stdin(Stdio::piped())
-                .stdout(Stdio::piped())
-                .stderr(Stdio::piped())
-                .env_remove("MYSQL_PWD")
-                .env_remove("MYSQL_PASSWORD")
-                .kill_on_drop(true);
-            let mut child = command.spawn().map_err(|error| {
-                if error.kind() == io::ErrorKind::NotFound {
-                    DbOperationError::CommandNotFound {
-                        command: DatabaseCli::MySql,
-                        details: error.to_string(),
-                    }
-                } else {
-                    DbOperationError::ConnectionFailed(error.to_string())
-                }
-            })?;
-            let stdin = child.stdin.take().ok_or_else(|| {
-                DbOperationError::QueryFailed("mysql stdin was not piped".to_string())
-            })?;
-            let stdout = child.stdout.take().ok_or_else(|| {
-                DbOperationError::QueryFailed("mysql stdout was not piped".to_string())
-            })?;
-            let stderr = child.stderr.take().ok_or_else(|| {
-                DbOperationError::QueryFailed("mysql stderr was not piped".to_string())
-            })?;
-            return Ok(Self {
-                child,
-                stdin,
-                stdout,
-                stderr,
-                pending: Vec::new(),
-                pending_stderr: Vec::new(),
-                frame_scanner: MysqlResultsetFrameScanner::default(),
-            });
-        }
-    }
-
-    #[cfg(unix)]
-    fn spawn_with_pty(
-        program: &OsStr,
-        option_file: &std::path::Path,
-    ) -> Result<Self, DbOperationError> {
-        let (master, slave) = create_mysql_pty().map_err(|error| {
-            DbOperationError::ConnectionFailed(format!("Unable to create MySQL PTY: {error}"))
-        })?;
-        let mut command = Command::new(program);
-        command
-            .args(mysql_query_args(option_file))
-            .stdin(Stdio::from(slave.try_clone().map_err(|error| {
-                DbOperationError::ConnectionFailed(error.to_string())
-            })?))
-            .stdout(Stdio::from(slave.try_clone().map_err(|error| {
-                DbOperationError::ConnectionFailed(error.to_string())
-            })?))
-            .stderr(Stdio::from(slave))
-            .env_remove("MYSQL_PWD")
-            .env_remove("MYSQL_PASSWORD")
-            .kill_on_drop(true);
-        let child = command.spawn().map_err(|error| {
-            if error.kind() == io::ErrorKind::NotFound {
-                DbOperationError::CommandNotFound {
-                    command: DatabaseCli::MySql,
-                    details: error.to_string(),
-                }
-            } else {
-                DbOperationError::ConnectionFailed(error.to_string())
-            }
-        })?;
-        let output = TokioFile::from_std(
-            master
-                .try_clone()
-                .map_err(|error| DbOperationError::ConnectionFailed(error.to_string()))?,
-        );
-        let input = TokioFile::from_std(master);
-        Ok(Self {
-            child,
-            pty: MysqlPty {
-                input,
-                output,
-                pending: Vec::new(),
-                frame_scanner: MysqlResultsetFrameScanner::default(),
-            },
-        })
-    }
-}
-
-struct MysqlMetadataSession {
-    process: MysqlProcess,
-}
-
-impl MysqlMetadataSession {
-    fn spawn_with_program(
-        program: &OsStr,
-        option_file: &std::path::Path,
-    ) -> Result<Self, DbOperationError> {
-        Ok(Self {
-            process: MysqlProcess::spawn_with_program(program, option_file)?,
-        })
-    }
-
-    async fn probe(&mut self) -> Result<(), DbOperationError> {
-        let marker = Uuid::new_v4().simple().to_string();
-        let query =
-            format!("SELECT '{marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode");
-        let result = self.execute(&query).await?;
-        validate_mode_probe(&result, &marker)
-    }
-
-    async fn execute(&mut self, query: &str) -> Result<MysqlResultSet, DbOperationError> {
-        write_mysql_statement(&mut self.process, query).await?;
-        let xml = read_one_mysql_resultset(&mut self.process).await?;
-        parse_mysql_xml(&xml)
-    }
-
-    async fn finish(&mut self) -> Result<(), DbOperationError> {
-        #[cfg(not(unix))]
-        self.process
-            .stdin
-            .shutdown()
-            .await
-            .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-
-        #[cfg(unix)]
-        let tail = {
-            write_mysql_input(&mut self.process, b"\x04").await?;
-            read_pty_all(&mut self.process.pty)
-                .await
-                .map_err(|error| DbOperationError::QueryFailed(error.to_string()))?
-        };
-
-        #[cfg(not(unix))]
-        let (_stdout, stderr) = tokio::join!(
-            read_all(&mut self.process.stdout),
-            read_all(&mut self.process.stderr)
-        );
-        #[cfg(not(unix))]
-        let stderr = stderr.map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-
-        let status = self
-            .process
-            .child
-            .wait()
-            .await
-            .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-        #[cfg(unix)]
-        let error_bytes = tail.as_slice();
-        #[cfg(not(unix))]
-        let error_bytes = stderr.as_slice();
-        if has_mysql_cli_error(error_bytes) {
-            return Err(classify_mysql_query_failure(error_bytes));
-        }
-        if !status.success() {
-            return Err(classify_mysql_query_failure(error_bytes));
-        }
-        Ok(())
-    }
-
-    async fn cleanup(&mut self) {
-        cleanup_mysql_process(&mut self.process).await;
-    }
-}
-
-#[cfg(unix)]
-fn create_mysql_pty() -> io::Result<(std::fs::File, std::fs::File)> {
-    let mut master = -1;
-    let mut slave = -1;
-    let result = unsafe {
-        libc::openpty(
-            &raw mut master,
-            &raw mut slave,
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-        )
-    };
-    if result != 0 {
-        return Err(io::Error::last_os_error());
-    }
-
-    let slave_file = unsafe { std::fs::File::from_raw_fd(slave) };
-    let master_file = unsafe { std::fs::File::from_raw_fd(master) };
-    let mut termios = std::mem::MaybeUninit::<libc::termios>::uninit();
-    if unsafe { libc::tcgetattr(slave_file.as_raw_fd(), termios.as_mut_ptr()) } == 0 {
-        let mut termios = unsafe { termios.assume_init() };
-        termios.c_lflag &= !(libc::ECHO | libc::ECHONL);
-        termios.c_oflag &= !libc::OPOST;
-        let _ =
-            unsafe { libc::tcsetattr(slave_file.as_raw_fd(), libc::TCSANOW, &raw const termios) };
-    }
-    Ok((master_file, slave_file))
-}
-
-async fn run_mysql_adhoc(
-    option_file: &std::path::Path,
-    query: &str,
-    statements: &[MysqlStatement],
-    access_mode: AccessMode,
-) -> Result<MysqlExecutionResult, DbOperationError> {
-    run_mysql_adhoc_with_program_and_statements(
-        OsStr::new("mysql"),
-        option_file,
-        query,
-        statements,
-        access_mode,
-        MYSQL_QUERY_TIMEOUT,
-    )
-    .await
-}
-
-async fn export_mysql_csv_to_file(
-    target: MySqlDsn,
-    query: &str,
-    path: PathBuf,
-) -> Result<(), DbOperationError> {
-    let option_file = MySqlOptionFile::create(&target)?;
-    let mut process = MysqlProcess::spawn_with_program(OsStr::new("mysql"), &option_file.path)?;
-    let result = timeout(
-        MYSQL_EXPORT_TIMEOUT,
-        run_mysql_export_process(&mut process, &option_file.path, query, path),
-    )
-    .await;
-    match result {
-        Ok(Ok(())) => Ok(()),
-        Ok(Err(error)) => {
-            cleanup_mysql_process(&mut process).await;
-            Err(error)
-        }
-        Err(_) => {
-            cleanup_mysql_process(&mut process).await;
-            Err(DbOperationError::Timeout(
-                "mysql query exceeded the execution timeout".to_string(),
-            ))
-        }
-    }
-}
-
-async fn run_mysql_export_process(
-    process: &mut MysqlProcess,
-    option_file: &std::path::Path,
-    query: &str,
-    path: PathBuf,
-) -> Result<(), DbOperationError> {
-    let marker = Uuid::new_v4().simple().to_string();
-    let probe_query =
-        format!("SELECT '{marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode");
-    write_mysql_statement(process, &probe_query).await?;
-    let probe_xml = read_one_mysql_resultset(process).await?;
-    let probe = parse_mysql_xml(&probe_xml)?;
-    validate_mode_probe(&probe, &marker)?;
-    configure_mysql_session(process, AccessMode::ReadOnly).await?;
-
-    write_mysql_statement(process, query).await?;
-    let mut csv_writer = CsvFileWriter::create(path).await?;
-    let columns = stream_mysql_resultset_to_csv(process, &mut csv_writer).await?;
-    if columns.is_none() {
-        let statement = classify_mysql_statement(query)
-            .map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-        let fallback_kind = mysql_metadata_fallback_kind(&statement.kind).ok_or_else(|| {
-            DbOperationError::QueryFailed(
-                "MySQL empty CSV result has no supported metadata fallback".to_string(),
-            )
-        })?;
-        let columns = mysql_metadata_columns(process, option_file, query, fallback_kind).await?;
-        csv_writer.write_record(columns.iter()).await?;
-    }
-
-    #[cfg(not(unix))]
-    process
-        .stdin
-        .shutdown()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-
-    #[cfg(unix)]
-    let tail = {
-        write_mysql_input(process, b"\x04").await?;
-        read_pty_all(&mut process.pty)
-            .await
-            .map_err(|error| DbOperationError::QueryFailed(error.to_string()))?
-    };
-
-    #[cfg(not(unix))]
-    let (stdout, stderr) =
-        tokio::join!(read_all(&mut process.stdout), read_all(&mut process.stderr));
-    #[cfg(not(unix))]
-    let _stdout = stdout.map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-    #[cfg(not(unix))]
-    let stderr = stderr.map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-
-    let status = process
-        .child
-        .wait()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-    #[cfg(unix)]
-    let error_bytes = tail.as_slice();
-    #[cfg(not(unix))]
-    let error_bytes = stderr.as_slice();
-    if !status.success() {
-        return Err(classify_mysql_query_failure(error_bytes));
-    }
-
-    csv_writer.finish().await
-}
-
-async fn stream_mysql_resultset_to_csv(
-    process: &mut MysqlProcess,
-    csv_writer: &mut CsvFileWriter,
-) -> Result<Option<Vec<String>>, DbOperationError> {
-    #[cfg(unix)]
-    {
-        let source = MysqlExportPtySource {
-            pty: &mut process.pty,
-            error_output: Vec::new(),
-            pending: Vec::new(),
-            started: false,
-        };
-        let mut reader = Reader::from_reader(BufReader::new(source));
-        reader.config_mut().trim_text(false);
-        let result = stream_mysql_xml_to_csv(&mut reader, csv_writer).await;
-        let buffered = reader.into_inner();
-        let unread = buffered.buffer().to_vec();
-        let source = buffered.into_inner();
-        source.pty.pending.extend(unread);
-        source.pty.pending.extend(source.pending);
-        if !source.error_output.is_empty() {
-            return Err(classify_mysql_query_failure(&source.error_output));
-        }
-        result
-    }
-
-    #[cfg(not(unix))]
-    {
-        let source = MysqlExportPipeSource {
-            stdout: &mut process.stdout,
-            stderr: &mut process.stderr,
-            pending: &mut process.pending,
-            error_output: Vec::new(),
-            stderr_buffer: [0; 4096],
-            stderr_closed: false,
-            stdout_closed: false,
-        };
-        let mut reader = Reader::from_reader(BufReader::new(source));
-        reader.config_mut().trim_text(false);
-        let result = stream_mysql_xml_to_csv(&mut reader, csv_writer).await;
-        let buffered = reader.into_inner();
-        let unread = buffered.buffer().to_vec();
-        let source = buffered.into_inner();
-        source.pending.extend(unread);
-        if has_mysql_cli_error(&source.error_output) {
-            return Err(classify_mysql_query_failure(&source.error_output));
-        }
-        return result;
-    }
-}
-
-async fn stream_mysql_xml_to_csv<R>(
-    reader: &mut Reader<BufReader<R>>,
-    csv_writer: &mut CsvFileWriter,
-) -> Result<Option<Vec<String>>, DbOperationError>
-where
-    R: AsyncRead + Unpin,
-{
-    let mut buffer = Vec::new();
-    let mut resultset_count = 0;
-    let mut in_resultset = false;
-    let mut current_row: Option<Vec<(String, String)>> = None;
-    let mut current_field: Option<MysqlField> = None;
-    let mut columns: Option<Vec<String>> = None;
-
-    loop {
-        let event = reader
-            .read_event_into_async(&mut buffer)
-            .await
-            .map_err(|error| {
-                DbOperationError::QueryFailed(format!("invalid MySQL XML result: {error}"))
-            })?;
-        match event {
-            Event::Start(element) => match element.name().as_ref() {
-                b"resultset" => {
-                    if in_resultset || resultset_count > 0 {
-                        return Err(DbOperationError::QueryFailed(
-                            "mysql returned more than one resultset".to_string(),
-                        ));
-                    }
-                    resultset_count += 1;
-                    in_resultset = true;
-                }
-                b"row" if in_resultset && current_row.is_none() => {
-                    current_row = Some(Vec::new());
-                }
-                b"field" if current_row.is_some() && current_field.is_none() => {
-                    current_field = Some(parse_mysql_field(&element)?);
-                }
-                _ => {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected element in MySQL XML result".to_string(),
-                    ));
-                }
-            },
-            Event::Empty(element) if element.name().as_ref() == b"field" => {
-                let row = current_row.as_mut().ok_or_else(|| {
-                    DbOperationError::QueryFailed("MySQL XML field is outside a row".to_string())
-                })?;
-                if current_field.is_some() {
-                    return Err(DbOperationError::QueryFailed(
-                        "nested MySQL XML fields are not supported".to_string(),
-                    ));
-                }
-                row.push(parse_mysql_field(&element)?.finish_raw());
-            }
-            Event::Text(text) => {
-                let decoded = text.decode().map_err(|error| {
-                    DbOperationError::QueryFailed(format!("invalid MySQL XML text: {error}"))
-                })?;
-                let text = unescape(&decoded).map_err(|error| {
-                    DbOperationError::QueryFailed(format!("invalid MySQL XML text: {error}"))
-                })?;
-                if let Some(field) = current_field.as_mut() {
-                    field.value.push_str(&text);
-                } else if !text.chars().all(char::is_whitespace) {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected text in MySQL XML result".to_string(),
-                    ));
-                }
-            }
-            Event::GeneralRef(reference) => {
-                let text = decode_mysql_xml_reference(&reference)?;
-                if let Some(field) = current_field.as_mut() {
-                    field.value.push_str(&text);
-                } else if !text.chars().all(char::is_whitespace) {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected text in MySQL XML result".to_string(),
-                    ));
-                }
-            }
-            Event::CData(data) => {
-                if let Some(field) = current_field.as_mut() {
-                    field
-                        .value
-                        .push_str(std::str::from_utf8(data.as_ref()).map_err(|error| {
-                            DbOperationError::QueryFailed(format!(
-                                "invalid MySQL XML text: {error}"
-                            ))
-                        })?);
-                } else {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected CDATA in MySQL XML result".to_string(),
-                    ));
-                }
-            }
-            Event::End(element) => match element.name().as_ref() {
-                b"field" => {
-                    let row = current_row.as_mut().ok_or_else(|| {
-                        DbOperationError::QueryFailed(
-                            "MySQL XML field is outside a row".to_string(),
-                        )
-                    })?;
-                    let field = current_field.take().ok_or_else(|| {
-                        DbOperationError::QueryFailed("unexpected MySQL XML field end".to_string())
-                    })?;
-                    row.push(field.finish_raw());
-                }
-                b"row" => {
-                    let row = current_row.take().ok_or_else(|| {
-                        DbOperationError::QueryFailed("unexpected MySQL XML row end".to_string())
-                    })?;
-                    let row_columns = row.iter().map(|(name, _)| name.clone()).collect::<Vec<_>>();
-                    if let Some(columns) = columns.as_ref() {
-                        if row_columns != *columns {
-                            return Err(DbOperationError::QueryFailed(
-                                "MySQL XML rows have inconsistent fields".to_string(),
-                            ));
-                        }
-                    } else {
-                        csv_writer.write_record(row_columns.iter()).await?;
-                        columns = Some(row_columns);
-                    }
-                    csv_writer
-                        .write_record(row.iter().map(|(_, value)| value))
-                        .await?;
-                }
-                b"resultset" => {
-                    if !in_resultset || current_row.is_some() || current_field.is_some() {
-                        return Err(DbOperationError::QueryFailed(
-                            "malformed MySQL XML resultset".to_string(),
-                        ));
-                    }
-                    return Ok(columns);
-                }
-                _ => {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected MySQL XML closing element".to_string(),
-                    ));
-                }
-            },
-            Event::Decl(_) | Event::Comment(_) => {}
-            Event::Eof => break,
-            _ => {
-                return Err(DbOperationError::QueryFailed(
-                    "unexpected event in MySQL XML result".to_string(),
-                ));
-            }
-        }
-        buffer.clear();
-    }
-
-    if resultset_count != 1 || in_resultset || current_row.is_some() || current_field.is_some() {
-        return Err(DbOperationError::QueryFailed(
-            "MySQL XML result did not contain one complete resultset".to_string(),
-        ));
-    }
-    Ok(columns)
-}
-
-#[cfg(unix)]
-struct MysqlExportPtySource<'a> {
-    pty: &'a mut MysqlPty,
-    error_output: Vec<u8>,
-    pending: Vec<u8>,
-    started: bool,
-}
-
-#[cfg(unix)]
-impl MysqlExportPtySource<'_> {
-    fn capture_error(&mut self, bytes: &[u8]) {
-        if self.error_output.is_empty() && has_mysql_cli_error(bytes) {
-            self.error_output
-                .extend_from_slice(&bytes[..bytes.len().min(32 * 1024)]);
-        }
-    }
-
-    fn discard_before_resultset(&mut self) {
-        const RESULTSET_START: &[u8] = b"<resultset";
-        if let Some(start) = find_bytes(&self.pending, RESULTSET_START) {
-            self.pending.drain(..start);
-            self.started = true;
-        } else {
-            let keep = RESULTSET_START.len().saturating_sub(1);
-            let discard = self.pending.len().saturating_sub(keep);
-            self.pending.drain(..discard);
-        }
-    }
-}
-
-#[cfg(unix)]
-impl AsyncRead for MysqlExportPtySource<'_> {
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buffer: &mut ReadBuf<'_>,
-    ) -> Poll<io::Result<()>> {
-        let this = self.get_mut();
-        loop {
-            if !this.started {
-                if !this.pty.pending.is_empty() {
-                    let bytes = std::mem::take(&mut this.pty.pending);
-                    this.capture_error(&bytes);
-                    this.pending.extend_from_slice(&bytes);
-                }
-                if !this.pending.is_empty() {
-                    this.discard_before_resultset();
-                    if this.started {
-                        continue;
-                    }
-                }
-
-                let mut chunk = [0; 4096];
-                let mut read_buffer = ReadBuf::new(&mut chunk);
-                match Pin::new(&mut this.pty.output).poll_read(cx, &mut read_buffer) {
-                    Poll::Ready(Ok(())) => {
-                        let bytes = read_buffer.filled().to_vec();
-                        if bytes.is_empty() {
-                            return Poll::Ready(Ok(()));
-                        }
-                        this.capture_error(&bytes);
-                        this.pending.extend_from_slice(&bytes);
-                    }
-                    Poll::Ready(Err(error)) => return Poll::Ready(Err(error)),
-                    Poll::Pending => return Poll::Pending,
-                }
-                continue;
-            }
-
-            if !this.pending.is_empty() {
-                let count = buffer.remaining().min(this.pending.len());
-                let bytes = this.pending.drain(..count).collect::<Vec<_>>();
-                buffer.put_slice(&bytes);
-                return Poll::Ready(Ok(()));
-            }
-
-            {
-                let filled_before = buffer.filled().len();
-                let result = Pin::new(&mut this.pty.output).poll_read(cx, buffer);
-                if matches!(&result, Poll::Ready(Ok(()))) {
-                    this.capture_error(&buffer.filled()[filled_before..]);
-                }
-                return result;
-            }
-        }
-    }
-}
-
-#[cfg(not(unix))]
-struct MysqlExportPipeSource<'a, O, E> {
-    stdout: &'a mut O,
-    stderr: &'a mut E,
-    pending: &'a mut Vec<u8>,
-    error_output: Vec<u8>,
-    stderr_buffer: [u8; 4096],
-    stderr_closed: bool,
-    stdout_closed: bool,
-}
-
-#[cfg(not(unix))]
-impl<O, E> MysqlExportPipeSource<'_, O, E>
-where
-    O: AsyncRead + Unpin,
-    E: AsyncRead + Unpin,
-{
-    fn capture_error(&mut self, bytes: &[u8]) {
-        let remaining = (32usize * 1024).saturating_sub(self.error_output.len());
-        if remaining > 0 {
-            self.error_output
-                .extend_from_slice(&bytes[..bytes.len().min(remaining)]);
-        }
-    }
-
-    fn poll_stderr(&mut self, cx: &mut Context<'_>) -> Poll<io::Result<usize>> {
-        if self.stderr_closed {
-            return Poll::Ready(Ok(0));
-        }
-        let result = {
-            let mut read_buffer = ReadBuf::new(&mut self.stderr_buffer);
-            match Pin::new(&mut *self.stderr).poll_read(cx, &mut read_buffer) {
-                Poll::Ready(Ok(())) => Poll::Ready(Ok(read_buffer.filled().to_vec())),
-                Poll::Ready(Err(error)) => Poll::Ready(Err(error)),
-                Poll::Pending => Poll::Pending,
-            }
-        };
-        match result {
-            Poll::Ready(Ok(bytes)) => {
-                if bytes.is_empty() {
-                    self.stderr_closed = true;
-                } else {
-                    self.capture_error(&bytes);
-                }
-                Poll::Ready(Ok(bytes.len()))
-            }
-            Poll::Ready(Err(error)) => Poll::Ready(Err(error)),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-#[cfg(not(unix))]
-impl<O, E> AsyncRead for MysqlExportPipeSource<'_, O, E>
-where
-    O: AsyncRead + Unpin,
-    E: AsyncRead + Unpin,
-{
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buffer: &mut ReadBuf<'_>,
-    ) -> Poll<io::Result<()>> {
-        let this = self.get_mut();
-        let stderr_count = match this.poll_stderr(cx) {
-            Poll::Ready(Err(error)) => return Poll::Ready(Err(error)),
-            Poll::Pending => 0,
-            Poll::Ready(Ok(count)) => count,
-        };
-        if !this.pending.is_empty() {
-            let count = buffer.remaining().min(this.pending.len());
-            let bytes = this.pending.drain(..count).collect::<Vec<_>>();
-            buffer.put_slice(&bytes);
-            return Poll::Ready(Ok(()));
-        }
-
-        if this.stdout_closed {
-            if this.stderr_closed {
-                return Poll::Ready(Ok(()));
-            }
-            cx.waker().wake_by_ref();
-            return Poll::Pending;
-        }
-
-        let filled_before = buffer.filled().len();
-        let result = Pin::new(&mut *this.stdout).poll_read(cx, buffer);
-        if let Poll::Ready(Ok(())) = &result {
-            let count = buffer.filled().len() - filled_before;
-            if count == 0 {
-                this.stdout_closed = true;
-                if !this.stderr_closed {
-                    cx.waker().wake_by_ref();
-                    return Poll::Pending;
-                }
-            }
-        } else if matches!(&result, Poll::Pending) && stderr_count > 0 {
-            cx.waker().wake_by_ref();
-        }
-        result
-    }
-}
-
-#[cfg(all(test, not(unix)))]
-mod export_pipe_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn consumes_stderr_while_streaming_stdout() {
-        let (mut stdout_writer, mut stdout_reader) = tokio::io::duplex(64);
-        let (mut stderr_writer, mut stderr_reader) = tokio::io::duplex(64);
-        let stdout = b"<resultset><row><field name=\"value\">ok</field></row></resultset>".to_vec();
-        let stderr = format!(
-            "ERROR 1146 (42S02): Table 'app.missing' doesn't exist\n{}",
-            "warning\n".repeat(16 * 1024)
-        )
-        .into_bytes();
-
-        let stdout_task = tokio::spawn(async move {
-            stdout_writer.write_all(&stdout).await.unwrap();
-        });
-        let stderr_task = tokio::spawn(async move {
-            stderr_writer.write_all(&stderr).await.unwrap();
-        });
-
-        let mut source = MysqlExportPipeSource {
-            stdout: &mut stdout_reader,
-            stderr: &mut stderr_reader,
-            pending: &mut Vec::new(),
-            error_output: Vec::new(),
-            stderr_buffer: [0; 4096],
-            stderr_closed: false,
-            stdout_closed: false,
-        };
-        let mut output = Vec::new();
-        source.read_to_end(&mut output).await.unwrap();
-        stdout_task.await.unwrap();
-        stderr_task.await.unwrap();
-
-        assert!(output.starts_with(b"<resultset>"));
-        assert!(has_mysql_cli_error(&source.error_output));
-        assert!(matches!(
-            classify_mysql_query_failure(&source.error_output),
-            DbOperationError::ObjectMissing(_)
-        ));
-    }
-}
-
-async fn run_mysql_single_statement(
-    option_file: &std::path::Path,
-    query: &str,
-    access_mode: AccessMode,
-) -> Result<MysqlResultSet, DbOperationError> {
-    let mut process = MysqlProcess::spawn_with_program(OsStr::new("mysql"), option_file)?;
-    let result = timeout(
-        MYSQL_QUERY_TIMEOUT,
-        run_mysql_single_statement_process(&mut process, query, access_mode),
-    )
-    .await;
-    match result {
-        Ok(Ok(result_set)) => Ok(result_set),
-        Ok(Err(error)) => {
-            cleanup_mysql_process(&mut process).await;
-            Err(error)
-        }
-        Err(_) => {
-            cleanup_mysql_process(&mut process).await;
-            Err(DbOperationError::Timeout(
-                "mysql query exceeded the execution timeout".to_string(),
-            ))
-        }
-    }
-}
-
-async fn run_mysql_single_statement_process(
-    process: &mut MysqlProcess,
-    query: &str,
-    access_mode: AccessMode,
-) -> Result<MysqlResultSet, DbOperationError> {
-    let marker = Uuid::new_v4().simple().to_string();
-    let probe_query =
-        format!("SELECT '{marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode");
-    write_mysql_statement(process, &probe_query).await?;
-    let probe_xml = read_one_mysql_resultset(process).await?;
-    let probe = parse_mysql_xml(&probe_xml)?;
-    validate_mode_probe(&probe, &marker)?;
-    configure_mysql_session(process, access_mode).await?;
-
-    write_mysql_statement(process, query).await?;
-
-    #[cfg(not(unix))]
-    process
-        .stdin
-        .shutdown()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-
-    #[cfg(unix)]
-    let (stdout, tail) = {
-        let stdout = read_one_mysql_resultset(process).await?;
-        write_mysql_input(process, b"\x04").await?;
-        let tail = read_pty_all(&mut process.pty)
-            .await
-            .map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-        (stdout, tail)
-    };
-
-    #[cfg(not(unix))]
-    let (stdout, stderr) =
-        tokio::join!(read_all(&mut process.stdout), read_all(&mut process.stderr));
-    #[cfg(not(unix))]
-    let stdout = stdout.map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-    #[cfg(not(unix))]
-    let stderr = stderr.map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-
-    let status = process
-        .child
-        .wait()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-    #[cfg(unix)]
-    let error_bytes = tail.as_slice();
-    #[cfg(not(unix))]
-    let error_bytes = stderr.as_slice();
-    if !status.success() {
-        return Err(classify_mysql_query_failure(error_bytes));
-    }
-    parse_mysql_xml(&stdout)
-}
-
-async fn run_mysql_adhoc_with_program_and_statements(
-    program: &OsStr,
-    option_file: &std::path::Path,
-    query: &str,
-    statements: &[MysqlStatement],
-    access_mode: AccessMode,
-    execution_timeout: Duration,
-) -> Result<MysqlExecutionResult, DbOperationError> {
-    if mysql_metadata_fallback_has_unsupported_session_state(statements) {
-        return Err(DbOperationError::UnsupportedOperation(
-            "MySQL empty SHOW/DESCRIBE metadata fallback cannot preserve temporary-table session state"
-                .to_string(),
-        ));
-    }
-    let mut process = MysqlProcess::spawn_with_program(program, option_file)?;
-    let result = timeout(
-        execution_timeout,
-        run_mysql_adhoc_process(&mut process, option_file, query, statements, access_mode),
-    )
-    .await;
-
-    match result {
-        Ok(Ok(result_set)) => Ok(result_set),
-        Ok(Err(error)) => {
-            cleanup_mysql_process(&mut process).await;
-            Err(error)
-        }
-        Err(_) => {
-            cleanup_mysql_process(&mut process).await;
-            Err(DbOperationError::Timeout(
-                "mysql query exceeded the execution timeout".to_string(),
-            ))
-        }
-    }
-}
-
-async fn run_mysql_adhoc_process(
-    process: &mut MysqlProcess,
-    option_file: &std::path::Path,
-    _query: &str,
-    statements: &[MysqlStatement],
-    access_mode: AccessMode,
-) -> Result<MysqlExecutionResult, DbOperationError> {
-    let probe_marker = Uuid::new_v4().simple().to_string();
-    let probe_query = format!(
-        "SELECT '{probe_marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode"
-    );
-    write_mysql_statement(process, &probe_query).await?;
-    let probe_xml = read_one_mysql_resultset(process).await?;
-    let probe = parse_mysql_xml(&probe_xml)?;
-    validate_mode_probe(&probe, &probe_marker)?;
-    configure_mysql_session(process, access_mode).await?;
-
-    let mut last_result_set = None;
-    let mut command_tags = Vec::with_capacity(statements.len());
-    let mut refresh_scope = RefreshScope::None;
-    let mut scope_before_statement = RefreshScope::None;
-
-    for statement in statements {
-        scope_before_statement = refresh_scope;
-        let marker = Uuid::new_v4().simple().to_string();
-        let statement_scope = mysql_refresh_scope(&statement.kind);
-        let possible_refresh_scope = refresh_scope.merge(statement_scope);
-        if let Err(error) = write_mysql_statement(process, &statement.sql).await {
-            return Err(query_failed_after_change(error, refresh_scope));
-        }
-        let marker_query =
-            format!("SELECT '{marker}' AS __sabiql_marker, ROW_COUNT() AS affected_rows");
-        if let Err(error) = write_mysql_statement(process, &marker_query).await {
-            return Err(query_failed_after_change(error, possible_refresh_scope));
-        }
-        let first_xml = match read_one_mysql_resultset(process).await {
-            Ok(xml) => xml,
-            Err(error) => {
-                return Err(query_failed_after_mysql_statement(
-                    error,
-                    refresh_scope,
-                    possible_refresh_scope,
-                ));
-            }
-        };
-        let first_result = match parse_mysql_xml(&first_xml) {
-            Ok(result) => result,
-            Err(error) => return Err(query_failed_after_change(error, possible_refresh_scope)),
-        };
-        let (user_result, marker_result) = if is_mysql_row_count_marker(&first_result, &marker) {
-            (None, first_result)
-        } else {
-            let xml = match read_one_mysql_resultset(process).await {
-                Ok(xml) => xml,
-                Err(error) => {
-                    return Err(query_failed_after_mysql_statement(
-                        error,
-                        refresh_scope,
-                        possible_refresh_scope,
-                    ));
-                }
-            };
-            let marker_result = match parse_mysql_xml(&xml) {
-                Ok(result) => result,
-                Err(error) => return Err(query_failed_after_change(error, possible_refresh_scope)),
-            };
-            let user_result = fill_mysql_empty_result_columns(
-                process,
-                first_result,
-                option_file,
-                &statement.sql,
-                &statement.kind,
-            )
-            .await
-            .map_err(|error| query_failed_after_change(error, possible_refresh_scope))?;
-            (Some(user_result), marker_result)
-        };
-        let affected_rows = match mysql_row_count_marker(&marker_result, &marker) {
-            Ok(rows) => rows,
-            Err(error) => return Err(query_failed_after_change(error, possible_refresh_scope)),
-        };
-        if let Some(result) = user_result {
-            last_result_set = Some(result);
-        }
-        let tag = mysql_command_tag(&statement.kind, affected_rows, last_result_set.as_ref());
-        command_tags.push(MysqlCommandEvent {
-            kind: statement.kind.clone(),
-            target: statement.target.clone(),
-            tag,
-        });
-        refresh_scope = possible_refresh_scope;
-    }
-
-    #[cfg(not(unix))]
-    process
-        .stdin
-        .shutdown()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-
-    #[cfg(unix)]
-    let tail = {
-        write_mysql_input(process, b"\x04").await?;
-        let tail = read_pty_all(&mut process.pty)
-            .await
-            .map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-        trace_mysql_frame("discard tail", tail.len());
-        trace_mysql_error(&tail);
-        tail
-    };
-
-    #[cfg(not(unix))]
-    let (_stdout, stderr) =
-        tokio::join!(read_all(&mut process.stdout), read_all(&mut process.stderr));
-    #[cfg(not(unix))]
-    let stderr = stderr.map_err(|error| DbOperationError::QueryFailed(error.to_string()))?;
-
-    let status = process
-        .child
-        .wait()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-
-    #[cfg(unix)]
-    let error_bytes = tail.as_slice();
-    #[cfg(not(unix))]
-    let error_bytes = stderr.as_slice();
-    if has_mysql_cli_error(error_bytes) {
-        return Err(query_failed_after_change(
-            classify_mysql_query_failure(error_bytes),
-            scope_before_statement,
-        ));
-    }
-    if !status.success() {
-        return Err(query_failed_after_change(
-            classify_mysql_query_failure(error_bytes),
-            refresh_scope,
-        ));
-    }
-
-    Ok(MysqlExecutionResult {
-        result_set: last_result_set,
-        command_tag: aggregate_mysql_command_tag(&command_tags),
-        refresh_scope,
-    })
-}
-
-async fn configure_mysql_session(
-    process: &mut MysqlProcess,
-    access_mode: AccessMode,
-) -> Result<(), DbOperationError> {
-    if !access_mode.is_read_only() {
-        return Ok(());
-    }
-
-    let marker = Uuid::new_v4().simple().to_string();
-    write_mysql_statement(process, MYSQL_READ_ONLY_STATEMENT).await?;
-    write_mysql_statement(
-        process,
-        &format!("SELECT '{marker}' AS {MYSQL_SESSION_MARKER_COLUMN}"),
-    )
-    .await?;
-    loop {
-        let result = read_one_mysql_resultset(process).await?;
-        let result = parse_mysql_xml(&result)?;
-        if result.columns.is_empty() && result.values.is_empty() {
-            continue;
-        }
-        return validate_mysql_session_marker(&result, &marker);
-    }
-}
-
-fn validate_mysql_session_marker(
-    result: &MysqlResultSet,
-    marker: &str,
-) -> Result<(), DbOperationError> {
-    if result.columns != [MYSQL_SESSION_MARKER_COLUMN]
-        || result.values.len() != 1
-        || result.values[0].len() != 1
-        || result.values[0][0].as_str() != Some(marker)
-    {
-        return Err(DbOperationError::QueryFailed(
-            "mysql read-only session marker did not match".to_string(),
-        ));
-    }
-    Ok(())
-}
-
-fn query_failed_after_change(
-    error: DbOperationError,
-    refresh_scope: RefreshScope,
-) -> DbOperationError {
-    if refresh_scope == RefreshScope::None {
-        error
-    } else {
-        DbOperationError::QueryFailedAfterChange {
-            source: Arc::new(error),
-            refresh_scope,
-        }
-    }
-}
-
-fn query_failed_after_mysql_statement(
-    error: DbOperationError,
-    refresh_scope: RefreshScope,
-    possible_refresh_scope: RefreshScope,
-) -> DbOperationError {
-    let refresh_scope = if is_mysql_statement_failure(&error) {
-        refresh_scope
-    } else {
-        possible_refresh_scope
-    };
-    query_failed_after_change(error, refresh_scope)
-}
-
-fn is_mysql_statement_failure(error: &DbOperationError) -> bool {
-    matches!(
-        error,
-        DbOperationError::PermissionDenied(_)
-            | DbOperationError::ForeignKeyViolation(_)
-            | DbOperationError::UniqueViolation(_)
-            | DbOperationError::LockTimeout(_)
-            | DbOperationError::ObjectMissing(_)
-            | DbOperationError::QueryFailed(_)
-            | DbOperationError::Canceled(_)
-    )
-}
-
-fn is_mysql_row_count_marker(result: &MysqlResultSet, marker: &str) -> bool {
-    result.columns == ["__sabiql_marker", "affected_rows"]
-        && result.values.len() == 1
-        && result.values[0].first().and_then(QueryValue::as_str) == Some(marker)
-}
-
-fn mysql_row_count_marker(result: &MysqlResultSet, marker: &str) -> Result<i64, DbOperationError> {
-    if !is_mysql_row_count_marker(result, marker) || result.values[0].len() != 2 {
-        return Err(DbOperationError::QueryFailed(
-            "mysql ROW_COUNT marker did not match the executed statement".to_string(),
-        ));
-    }
-    let value = result.values[0][1].as_str().ok_or_else(|| {
-        DbOperationError::QueryFailed("mysql ROW_COUNT marker was NULL".to_string())
-    })?;
-    value.parse::<i64>().map_err(|_| {
-        DbOperationError::QueryFailed("mysql ROW_COUNT marker was not an integer".to_string())
-    })
-}
-
-fn mysql_command_tag(
-    kind: &MysqlStatementKind,
-    affected_rows: i64,
-    user_result: Option<&MysqlResultSet>,
-) -> CommandTag {
-    let rows = || u64::try_from(affected_rows.max(0)).unwrap_or(0);
-    match kind {
-        MysqlStatementKind::Select
-        | MysqlStatementKind::Table
-        | MysqlStatementKind::Show
-        | MysqlStatementKind::Describe => {
-            CommandTag::Select(user_result.map_or(0, |result| result.values.len() as u64))
-        }
-        MysqlStatementKind::Insert | MysqlStatementKind::Replace => CommandTag::Insert(rows()),
-        MysqlStatementKind::Update { .. } => CommandTag::Update(rows()),
-        MysqlStatementKind::Delete { .. } => CommandTag::Delete(rows()),
-        MysqlStatementKind::CreateTable { temporary: true } => {
-            CommandTag::Other("CREATE TEMPORARY TABLE".to_string())
-        }
-        MysqlStatementKind::CreateTable { temporary: false } => {
-            CommandTag::Create("TABLE".to_string())
-        }
-        MysqlStatementKind::AlterTable => CommandTag::Alter("TABLE".to_string()),
-        MysqlStatementKind::DropTable { temporary: true } => {
-            CommandTag::Other("DROP TEMPORARY TABLE".to_string())
-        }
-        MysqlStatementKind::DropTable { temporary: false } => CommandTag::Drop("TABLE".to_string()),
-        MysqlStatementKind::TruncateTable => CommandTag::Truncate,
-        MysqlStatementKind::CreateView => CommandTag::Create("VIEW".to_string()),
-        MysqlStatementKind::DropView => CommandTag::Drop("VIEW".to_string()),
-        MysqlStatementKind::CreateIndex => CommandTag::Create("INDEX".to_string()),
-        MysqlStatementKind::DropIndex => CommandTag::Drop("INDEX".to_string()),
-        MysqlStatementKind::Begin | MysqlStatementKind::StartTransaction => CommandTag::Begin,
-        MysqlStatementKind::Commit => CommandTag::Commit,
-        MysqlStatementKind::Rollback | MysqlStatementKind::RollbackToSavepoint => {
-            CommandTag::Rollback
-        }
-        MysqlStatementKind::Savepoint => CommandTag::Other("SAVEPOINT".to_string()),
-        MysqlStatementKind::ReleaseSavepoint => CommandTag::Other("RELEASE SAVEPOINT".to_string()),
-    }
-}
-
-fn mysql_refresh_scope(kind: &MysqlStatementKind) -> RefreshScope {
-    if mysql_statement_is_schema_modifying(kind) {
-        RefreshScope::Metadata
-    } else if mysql_statement_is_data_modifying(kind) {
-        RefreshScope::Data
-    } else {
-        RefreshScope::None
-    }
-}
-
-fn mysql_statement_is_persistent_schema_change(kind: &MysqlStatementKind) -> bool {
-    mysql_statement_is_schema_modifying(kind)
-        && !matches!(
-            kind,
-            MysqlStatementKind::CreateTable { temporary: true }
-                | MysqlStatementKind::DropTable { temporary: true }
-        )
-}
-
-#[derive(Default)]
-struct MysqlPendingTransactionTags {
-    data: Vec<CommandTag>,
-    savepoints: Vec<(String, usize)>,
-}
-
-fn apply_pending_mysql_data(
-    pending: MysqlPendingTransactionTags,
-    committed_data: &mut Option<CommandTag>,
-) {
-    if let Some(tag) = pending.data.last() {
-        *committed_data = Some(tag.clone());
-    }
-}
-
-fn aggregate_mysql_command_tag(events: &[MysqlCommandEvent]) -> Option<CommandTag> {
-    let mut committed_schema = None;
-    let mut committed_data = None;
-    let mut pending = None;
-    let mut last_tag = None;
-
-    for event in events {
-        last_tag = Some(event.tag.clone());
-        match &event.kind {
-            MysqlStatementKind::Begin | MysqlStatementKind::StartTransaction => {
-                pending = Some(MysqlPendingTransactionTags::default());
-            }
-            MysqlStatementKind::Commit => {
-                if let Some(transaction) = pending.take() {
-                    apply_pending_mysql_data(transaction, &mut committed_data);
-                }
-            }
-            MysqlStatementKind::Rollback => {
-                pending = None;
-            }
-            MysqlStatementKind::Savepoint => {
-                if let Some(transaction) = pending.as_mut()
-                    && let Some(name) = event.target.as_deref()
-                {
-                    transaction
-                        .savepoints
-                        .retain(|(current, _)| !current.eq_ignore_ascii_case(name));
-                    transaction
-                        .savepoints
-                        .push((name.to_string(), transaction.data.len()));
-                }
-            }
-            MysqlStatementKind::RollbackToSavepoint => {
-                if let Some(transaction) = pending.as_mut()
-                    && let Some(name) = event.target.as_deref()
-                    && let Some(index) = transaction
-                        .savepoints
-                        .iter()
-                        .position(|(current, _)| current.eq_ignore_ascii_case(name))
-                {
-                    transaction.data.truncate(transaction.savepoints[index].1);
-                    transaction.savepoints.truncate(index + 1);
-                }
-            }
-            MysqlStatementKind::ReleaseSavepoint => {
-                if let Some(transaction) = pending.as_mut()
-                    && let Some(name) = event.target.as_deref()
-                    && let Some(index) = transaction
-                        .savepoints
-                        .iter()
-                        .position(|(current, _)| current.eq_ignore_ascii_case(name))
-                {
-                    transaction.savepoints.remove(index);
-                }
-            }
-            MysqlStatementKind::CreateTable { temporary: true }
-            | MysqlStatementKind::DropTable { temporary: true } => {}
-            kind if mysql_statement_is_persistent_schema_change(kind) => {
-                if let Some(transaction) = pending.take() {
-                    apply_pending_mysql_data(transaction, &mut committed_data);
-                }
-                committed_schema = Some(event.tag.clone());
-            }
-            kind if mysql_statement_is_data_modifying(kind) => {
-                if let Some(transaction) = pending.as_mut() {
-                    transaction.data.push(event.tag.clone());
-                } else {
-                    committed_data = Some(event.tag.clone());
-                }
-            }
-            _ => {}
-        }
-    }
-
-    committed_schema.or(committed_data).or(last_tag)
-}
-
-async fn write_mysql_statement(
-    process: &mut MysqlProcess,
-    query: &str,
-) -> Result<(), DbOperationError> {
-    let query = query.trim_end();
-    write_mysql_input(process, query.as_bytes()).await?;
-    if query.ends_with(';') {
-        write_mysql_input(process, b"\n").await
-    } else if mysql_statement_has_trailing_line_comment(query) {
-        write_mysql_input(process, b"\n;\n").await
-    } else {
-        write_mysql_input(process, b";\n").await
-    }
-}
-
-fn mysql_statement_has_trailing_line_comment(sql: &str) -> bool {
-    let bytes = sql.as_bytes();
-    let mut index = 0;
-    let mut quote = None;
-    let mut line_comment = false;
-    while index < bytes.len() {
-        if let Some(delimiter) = quote {
-            if bytes[index] == b'\\' && delimiter != b'`' {
-                index += 2;
-            } else if bytes[index] == delimiter {
-                if bytes.get(index + 1) == Some(&delimiter) {
-                    index += 2;
-                } else {
-                    quote = None;
-                    index += 1;
-                }
-            } else {
-                index += 1;
-            }
-            continue;
-        }
-        if line_comment {
-            if bytes[index] == b'\n' {
-                line_comment = false;
-            }
-            index += 1;
-            continue;
-        }
-        if matches!(bytes[index], b'\'' | b'"' | b'`') {
-            quote = Some(bytes[index]);
-            index += 1;
-        } else if mysql_is_line_comment_start(bytes, index) {
-            let comment_start = index;
-            index = mysql_skip_line_comment(bytes, index);
-            line_comment = !bytes[comment_start..index].contains(&b'\n');
-        } else if bytes.get(index..index + 2) == Some(b"/*") {
-            index = mysql_skip_block_comment(bytes, index);
-        } else {
-            index += 1;
-        }
-    }
-    line_comment
-}
-
-fn mysql_is_line_comment_start(bytes: &[u8], index: usize) -> bool {
-    bytes[index] == b'#'
-        || (bytes.get(index..index + 2) == Some(b"--")
-            && bytes.get(index + 2).is_none_or(u8::is_ascii_whitespace))
-}
-
-fn mysql_skip_line_comment(bytes: &[u8], mut index: usize) -> usize {
-    while index < bytes.len() {
-        let byte = bytes[index];
-        index += 1;
-        if byte == b'\n' {
-            break;
-        }
-    }
-    index
-}
-
-fn mysql_skip_block_comment(bytes: &[u8], index: usize) -> usize {
-    let mut cursor = index + 2;
-    while cursor + 1 < bytes.len() {
-        if bytes[cursor] == b'*' && bytes[cursor + 1] == b'/' {
-            return cursor + 2;
-        }
-        cursor += 1;
-    }
-    bytes.len()
-}
-
-async fn write_mysql_input(
-    process: &mut MysqlProcess,
-    input: &[u8],
-) -> Result<(), DbOperationError> {
-    #[cfg(unix)]
-    process
-        .pty
-        .input
-        .write_all(input)
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-    #[cfg(not(unix))]
-    process
-        .stdin
-        .write_all(input)
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-    #[cfg(unix)]
-    process
-        .pty
-        .input
-        .flush()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-    #[cfg(not(unix))]
-    process
-        .stdin
-        .flush()
-        .await
-        .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-    Ok(())
-}
-
-async fn cleanup_mysql_process(process: &mut MysqlProcess) {
-    let _ = process.child.kill().await;
-    #[cfg(unix)]
-    let _ = read_pty_all(&mut process.pty).await;
-    #[cfg(not(unix))]
-    let _ = tokio::join!(read_all(&mut process.stdout), read_all(&mut process.stderr));
-    let _ = process.child.wait().await;
-}
-
-async fn read_one_mysql_resultset(process: &mut MysqlProcess) -> Result<Vec<u8>, DbOperationError> {
-    #[cfg(unix)]
-    {
-        return read_one_pty_resultset(&mut process.pty).await;
-    }
-    #[cfg(not(unix))]
-    read_one_mysql_resultset_from_pipes(
-        &mut process.stdout,
-        &mut process.stderr,
-        &mut process.pending,
-        &mut process.pending_stderr,
-        &mut process.frame_scanner,
-    )
-    .await
-}
-
-#[cfg(unix)]
-async fn read_one_pty_resultset(pty: &mut MysqlPty) -> Result<Vec<u8>, DbOperationError> {
-    let mut chunk = [0; 4096];
-    loop {
-        if let Some(frame) =
-            take_mysql_pty_resultset_frame(&mut pty.pending, &mut pty.frame_scanner)?
-        {
-            trace_mysql_frame("receive resultset", frame.len());
-            return Ok(frame);
-        }
-        let count = match pty.output.read(&mut chunk).await {
-            Ok(count) => count,
-            Err(error) if error.raw_os_error() == Some(libc::EIO) => 0,
-            Err(error) => return Err(DbOperationError::ConnectionLost(error.to_string())),
-        };
-        if count == 0 {
-            let tail = read_pty_all(pty)
-                .await
-                .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-            if has_mysql_cli_error(&tail) {
-                trace_mysql_error(&tail);
-                return Err(classify_mysql_query_failure(&tail));
-            }
-            return Err(DbOperationError::EmptyResponse(
-                "mysql query returned no resultset".to_string(),
-            ));
-        }
-        pty.pending.extend_from_slice(&chunk[..count]);
-    }
-}
-
-#[cfg(unix)]
-async fn read_pty_all(pty: &mut MysqlPty) -> io::Result<Vec<u8>> {
-    let mut output = std::mem::take(&mut pty.pending);
-    pty.frame_scanner.reset();
-    let mut chunk = [0; 4096];
-    loop {
-        match pty.output.read(&mut chunk).await {
-            Ok(0) => return Ok(output),
-            Ok(count) => output.extend_from_slice(&chunk[..count]),
-            Err(error) if matches!(error.raw_os_error(), Some(libc::EIO | libc::EPERM)) => {
-                return Ok(output);
-            }
-            Err(error) => return Err(error),
-        }
-    }
-}
-
-fn has_mysql_cli_error(output: &[u8]) -> bool {
-    output
-        .split(|byte| *byte == b'\n' || *byte == b'\r')
-        .any(|line| {
-            let mut line = line;
-            while line
-                .first()
-                .is_some_and(|byte| byte.is_ascii_whitespace() || byte.is_ascii_control())
-            {
-                line = &line[1..];
-            }
-            line.starts_with(b"ERROR ") || line == b"ERROR"
-        })
-}
-
-fn is_mysql_batch_diagnostic(line: &[u8]) -> bool {
-    line.starts_with(b"mysql: ") || line.starts_with(b"Warning: ")
-}
-
-#[cfg(any(unix, test))]
-fn take_mysql_pty_resultset_frame(
-    buffer: &mut Vec<u8>,
-    scanner: &mut MysqlResultsetFrameScanner,
-) -> Result<Option<Vec<u8>>, DbOperationError> {
-    let bounds = scanner.frame_bounds(buffer);
-    let resultset_start = scanner.resultset_start.unwrap_or(buffer.len());
-    if has_mysql_cli_error(&buffer[..resultset_start]) {
-        trace_mysql_error(&buffer[..resultset_start]);
-        return Err(classify_mysql_query_failure(&buffer[..resultset_start]));
-    }
-    Ok(bounds.map(|bounds| scanner.take_bounds(buffer, bounds)))
-}
-
-#[cfg(any(not(unix), test))]
-fn take_mysql_resultset_frame_after_error_check(
-    buffer: &mut Vec<u8>,
-    error_output: &[u8],
-    scanner: &mut MysqlResultsetFrameScanner,
-) -> Result<Option<Vec<u8>>, DbOperationError> {
-    if has_mysql_cli_error(error_output) {
-        trace_mysql_error(error_output);
-        return Err(classify_mysql_query_failure(error_output));
-    }
-    Ok(scanner.take(buffer))
-}
-
-#[cfg(unix)]
-fn find_bytes(haystack: &[u8], needle: &[u8]) -> Option<usize> {
-    find_bytes_from(haystack, needle, 0)
-}
-
-fn find_bytes_from(haystack: &[u8], needle: &[u8], start: usize) -> Option<usize> {
-    haystack
-        .get(start..)?
-        .windows(needle.len())
-        .position(|window| window == needle)
-        .map(|offset| start + offset)
-}
-
-fn trace_mysql_frame(kind: &str, bytes: usize) {
-    if std::env::var_os("SABIQL_MYSQL_TRANSCRIPT").is_some() {
-        write_mysql_transcript_line(&format!("sabiql mysql frame: {kind}, bytes={bytes}"));
-    }
-}
-
-fn trace_mysql_error(output: &[u8]) {
-    if std::env::var_os("SABIQL_MYSQL_TRANSCRIPT").is_some() && has_mysql_cli_error(output) {
-        write_mysql_transcript_line("sabiql mysql frame: ERROR line observed");
-    }
-}
-
-fn write_mysql_transcript_line(line: &str) {
-    let mut stderr = io::stderr();
-    let _ = stderr.write_all(line.as_bytes());
-    let _ = stderr.write_all(b"\n");
-}
-
-#[cfg(not(unix))]
-async fn read_all<R>(reader: &mut R) -> io::Result<Vec<u8>>
-where
-    R: AsyncRead + Unpin,
-{
-    let mut output = Vec::new();
-    reader.read_to_end(&mut output).await?;
-    Ok(output)
-}
-
-#[cfg(not(unix))]
-async fn read_one_mysql_resultset_from_pipes<R, E>(
-    reader: &mut R,
-    stderr: &mut E,
-    pending: &mut Vec<u8>,
-    pending_stderr: &mut Vec<u8>,
-    frame_scanner: &mut MysqlResultsetFrameScanner,
-) -> Result<Vec<u8>, DbOperationError>
-where
-    R: AsyncRead + Unpin,
-    E: AsyncRead + Unpin,
-{
-    let mut chunk = [0; 4096];
-    let mut stderr_chunk = [0; 4096];
-    let mut stderr_closed = false;
-    loop {
-        if frame_scanner.frame_bounds(pending).is_some() && !stderr_closed {
-            tokio::select! {
-                biased;
-                result = stderr.read(&mut stderr_chunk) => {
-                    let count = result.map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-                    if count == 0 {
-                        stderr_closed = true;
-                    } else {
-                        pending_stderr.extend_from_slice(&stderr_chunk[..count]);
-                    }
-                }
-                _ = tokio::task::yield_now() => {}
-            }
-        }
-        if let Some(frame) =
-            take_mysql_resultset_frame_after_error_check(pending, pending_stderr, frame_scanner)?
-        {
-            trace_mysql_frame("receive resultset", frame.len());
-            return Ok(frame);
-        }
-        if stderr_closed {
-            let count = reader
-                .read(&mut chunk)
-                .await
-                .map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-            if count == 0 {
-                return Err(DbOperationError::EmptyResponse(
-                    "mysql mode probe returned no resultset".to_string(),
-                ));
-            }
-            pending.extend_from_slice(&chunk[..count]);
-        } else {
-            tokio::select! {
-                result = reader.read(&mut chunk) => {
-                    let count = result.map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-                    if count == 0 {
-                        return Err(DbOperationError::EmptyResponse(
-                            "mysql mode probe returned no resultset".to_string(),
-                        ));
-                    }
-                    pending.extend_from_slice(&chunk[..count]);
-                }
-                result = stderr.read(&mut stderr_chunk) => {
-                    let count = result.map_err(|error| DbOperationError::ConnectionLost(error.to_string()))?;
-                    if count == 0 {
-                        stderr_closed = true;
-                    } else {
-                        pending_stderr.extend_from_slice(&stderr_chunk[..count]);
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn mysql_query_args(option_file: &std::path::Path) -> Vec<String> {
-    vec![
-        format!("--defaults-file={}", option_file.display()),
-        "--no-login-paths".to_string(),
-        "--protocol=TCP".to_string(),
-        "--connect-timeout=10".to_string(),
-        "--xml".to_string(),
-        "--binary-as-hex".to_string(),
-        "--binary-mode".to_string(),
-        "--unbuffered".to_string(),
-        "--skip-reconnect".to_string(),
-        "--default-character-set=utf8mb4".to_string(),
-        "--batch".to_string(),
-        "--silent".to_string(),
-        "--prompt=".to_string(),
-    ]
-}
-
-fn mysql_metadata_args(option_file: &std::path::Path) -> Vec<String> {
-    vec![
-        format!("--defaults-file={}", option_file.display()),
-        "--no-login-paths".to_string(),
-        "--protocol=TCP".to_string(),
-        "--connect-timeout=10".to_string(),
-        "--batch".to_string(),
-        "--column-names".to_string(),
-        "--column-type-info".to_string(),
-        "--binary-as-hex".to_string(),
-        "--binary-mode".to_string(),
-        "--unbuffered".to_string(),
-        "--skip-reconnect".to_string(),
-        "--default-character-set=utf8mb4".to_string(),
-        "--prompt=".to_string(),
-    ]
-}
-
-async fn mysql_metadata_columns(
-    process: &mut MysqlProcess,
-    option_file: &std::path::Path,
-    query: &str,
-    kind: MysqlMetadataFallbackKind,
-) -> Result<Vec<String>, DbOperationError> {
-    let query = match kind {
-        MysqlMetadataFallbackKind::Select => {
-            return mysql_metadata_select_columns(process, query).await;
-        }
-        MysqlMetadataFallbackKind::Show | MysqlMetadataFallbackKind::Describe => {
-            query.trim().trim_end_matches(';').trim_end().to_string()
-        }
-    };
-    mysql_metadata_columns_external(option_file, &query).await
-}
-
-async fn mysql_metadata_select_columns(
-    process: &mut MysqlProcess,
-    query: &str,
-) -> Result<Vec<String>, DbOperationError> {
-    let suffix = Uuid::new_v4().simple().to_string();
-    let source_alias = format!("__sabiql_metadata_source_{suffix}");
-    let marker_alias = format!("__sabiql_metadata_marker_{suffix}");
-    let query = mysql_metadata_select_query(query, &source_alias, &marker_alias)?;
-    write_mysql_statement(process, &query).await?;
-    let xml = match read_one_mysql_resultset(process).await {
-        Err(DbOperationError::QueryFailed(details))
-            if details
-                .to_ascii_lowercase()
-                .contains("duplicate column name") =>
-        {
-            return Err(DbOperationError::UnsupportedOperation(
-                "MySQL SELECT metadata fallback does not support duplicate column names"
-                    .to_string(),
-            ));
-        }
-        result => result?,
-    };
-    let result = parse_mysql_xml(&xml)?;
-    let row = result.values.first().ok_or_else(|| {
-        DbOperationError::QueryFailed(
-            "MySQL SELECT metadata fallback returned no synthetic row".to_string(),
-        )
-    })?;
-    if result.values.len() != 1
-        || result.columns.is_empty()
-        || row.len() != result.columns.len()
-        || row.iter().any(|value| !matches!(value, QueryValue::Null))
-    {
-        return Err(DbOperationError::QueryFailed(
-            "MySQL SELECT metadata fallback returned an invalid synthetic row".to_string(),
-        ));
-    }
-    Ok(result.columns)
-}
-
-async fn mysql_metadata_columns_external(
-    option_file: &std::path::Path,
-    query: &str,
-) -> Result<Vec<String>, DbOperationError> {
-    let mut args = mysql_metadata_args(option_file);
-    args.push(format!("--execute={query}"));
-    let option_file = option_file.to_path_buf();
-    let output = run_mysql_command(args, Some(&option_file)).await?;
-    if !output.status.success() {
-        return Err(classify_mysql_query_failure(&output.stderr));
-    }
-    parse_mysql_metadata_header(&output.stdout, query)
-}
-
-fn parse_mysql_metadata_header(
-    output: &[u8],
-    query: &str,
-) -> Result<Vec<String>, DbOperationError> {
-    let query = query.trim_end();
-    let query_with_semicolon = format!("{query};");
-    let lines = output
-        .split(|byte| *byte == b'\n')
-        .map(|line| line.strip_suffix(b"\r").unwrap_or(line))
-        .filter(|line| {
-            !line.is_empty()
-                && !is_mysql_batch_diagnostic(line)
-                && *line != query.as_bytes()
-                && *line != query_with_semicolon.as_bytes()
-        })
-        .collect::<Vec<_>>();
-    let header = lines.first().ok_or_else(|| {
-        DbOperationError::QueryFailed(
-            "MySQL metadata fallback returned no column header".to_string(),
-        )
-    })?;
-    let columns = header
-        .split(|byte| *byte == b'\t')
-        .map(|column| {
-            String::from_utf8(column.to_vec()).map_err(|error| {
-                DbOperationError::QueryFailed(format!(
-                    "invalid MySQL metadata fallback column name: {error}"
-                ))
-            })
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    if columns.is_empty() || columns.iter().any(String::is_empty) {
-        return Err(DbOperationError::QueryFailed(
-            "MySQL metadata fallback returned an invalid column header".to_string(),
-        ));
-    }
-    if lines.len() != 1 {
-        return Err(DbOperationError::QueryFailed(
-            "MySQL metadata fallback returned data instead of a header".to_string(),
-        ));
-    }
-    Ok(columns)
-}
-
-async fn fill_mysql_empty_result_columns(
-    process: &mut MysqlProcess,
-    mut result: MysqlResultSet,
-    option_file: &std::path::Path,
-    query: &str,
-    kind: &MysqlStatementKind,
-) -> Result<MysqlResultSet, DbOperationError> {
-    if !result.columns.is_empty() || !result.values.is_empty() {
-        return Ok(result);
-    }
-    let fallback_kind = mysql_metadata_fallback_kind(kind).ok_or_else(|| {
-        DbOperationError::QueryFailed(
-            "MySQL empty result has no supported metadata fallback".to_string(),
-        )
-    })?;
-    result.columns = mysql_metadata_columns(process, option_file, query, fallback_kind).await?;
-    Ok(result)
-}
-
-fn validate_mode_probe(result: &MysqlResultSet, marker: &str) -> Result<(), DbOperationError> {
-    if result.values.len() != 1 || result.columns != ["__sabiql_probe", "__sabiql_sql_mode"] {
-        return Err(DbOperationError::QueryFailed(
-            "mysql sql_mode probe returned an unexpected result".to_string(),
-        ));
-    }
-    let values = &result.values[0];
-    if values.len() != 2 {
-        return Err(DbOperationError::QueryFailed(
-            "mysql sql_mode probe returned an unexpected result".to_string(),
-        ));
-    }
-    if values[0].as_str() != Some(marker) {
-        return Err(DbOperationError::QueryFailed(
-            "mysql sql_mode probe marker did not match".to_string(),
-        ));
-    }
-    let mode = values[1].as_str().ok_or_else(|| {
-        DbOperationError::QueryFailed("mysql sql_mode probe returned no mode".to_string())
-    })?;
-    validate_sql_mode(mode)
-}
-
-fn classify_mysql_query_failure(stderr: &[u8]) -> DbOperationError {
-    let details = clean_mysql_stderr(stderr, "mysql query failed");
-    let lower = details.to_ascii_lowercase();
-    let error_code = mysql_server_error_code(&lower);
-    if is_mysql_tls_error(&lower) {
-        DbOperationError::ConnectionFailed(details)
-    } else if is_mysql_connect_timeout_message(&details)
-        || lower.contains("connect timeout")
-        || lower.contains("connection timed out")
-    {
-        DbOperationError::Timeout(details)
-    } else if matches!(error_code, Some(1044 | 1142 | 1143 | 1227))
-        || lower.contains("command denied")
-    {
-        DbOperationError::PermissionDenied(details)
-    } else if error_code == Some(1045)
-        || lower.contains("access denied")
-        || lower.contains("authentication")
-    {
-        DbOperationError::ConnectionFailed(details)
-    } else if lower.contains("lost connection") || lower.contains("server has gone away") {
-        DbOperationError::ConnectionLost(details)
-    } else if lower.contains("lock wait timeout") || lower.contains("deadlock found") {
-        DbOperationError::LockTimeout(details)
-    } else if matches!(error_code, Some(1215 | 1216 | 1217 | 1451 | 1452))
-        || lower.contains("foreign key constraint")
-    {
-        DbOperationError::ForeignKeyViolation(details)
-    } else if lower.contains("doesn't exist") || lower.contains("does not exist") {
-        DbOperationError::ObjectMissing(details)
-    } else if lower.contains("duplicate entry") {
-        DbOperationError::UniqueViolation(details)
-    } else if lower.contains("query execution was interrupted")
-        || lower.contains("query was interrupted")
-    {
-        DbOperationError::Canceled(details)
-    } else {
-        DbOperationError::QueryFailed(details)
-    }
-}
-
-fn is_mysql_tls_error(lowercase_details: &str) -> bool {
-    [
-        "error 2026",
-        "tls/ssl error",
-        "ssl connection error",
-        "ssl handshake",
-        "tls handshake",
-        "handshake failure",
-        "tlsv1 alert",
-        "certificate verify failed",
-        "certificate verification failure",
-        "certificate validation failure",
-        "unable to get local issuer",
-        "self-signed certificate",
-        "unknown ca",
-        "certificate required",
-        "peer did not return a certificate",
-    ]
-    .iter()
-    .any(|marker| lowercase_details.contains(marker))
-}
-
-fn mysql_server_error_code(lowercase_details: &str) -> Option<u32> {
-    let marker = "error ";
-    let start = lowercase_details.find(marker)? + marker.len();
-    let digits = &lowercase_details[start..];
-    let end = digits
-        .find(|character: char| !character.is_ascii_digit())
-        .unwrap_or(digits.len());
-    digits[..end].parse().ok()
-}
-
-fn clean_mysql_stderr(stderr: &[u8], fallback: &str) -> String {
-    let text = String::from_utf8_lossy(stderr).trim().to_string();
-    if text.is_empty() {
-        fallback.to_string()
-    } else {
-        text
-    }
-}
-
-fn decode_mysql_xml_reference(reference: &BytesRef<'_>) -> Result<String, DbOperationError> {
-    let reference = reference.decode().map_err(|error| {
-        DbOperationError::QueryFailed(format!("invalid MySQL XML text: {error}"))
-    })?;
-    unescape(&format!("&{reference};"))
-        .map(std::borrow::Cow::into_owned)
-        .map_err(|error| DbOperationError::QueryFailed(format!("invalid MySQL XML text: {error}")))
-}
-
-fn parse_mysql_xml(xml: &[u8]) -> Result<MysqlResultSet, DbOperationError> {
-    let mut reader = Reader::from_reader(xml);
-    reader.config_mut().trim_text(false);
-    let mut buffer = Vec::new();
-    let mut resultset_count = 0;
-    let mut in_resultset = false;
-    let mut current_row: Option<Vec<(String, QueryValue)>> = None;
-    let mut current_field: Option<MysqlField> = None;
-    let mut rows = Vec::new();
-    let mut columns = Vec::new();
-
-    loop {
-        let event = reader.read_event_into(&mut buffer).map_err(|error| {
-            DbOperationError::QueryFailed(format!("invalid MySQL XML result: {error}"))
-        })?;
-        match event {
-            Event::Start(element) => match element.name().as_ref() {
-                b"resultset" => {
-                    if in_resultset || resultset_count > 0 {
-                        return Err(DbOperationError::QueryFailed(
-                            "mysql returned more than one resultset".to_string(),
-                        ));
-                    }
-                    resultset_count += 1;
-                    in_resultset = true;
-                }
-                b"row" if in_resultset && current_row.is_none() => {
-                    current_row = Some(Vec::new());
-                }
-                b"field" if current_row.is_some() && current_field.is_none() => {
-                    current_field = Some(parse_mysql_field(&element)?);
-                }
-                _ => {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected element in MySQL XML result".to_string(),
-                    ));
-                }
-            },
-            Event::Empty(element) if element.name().as_ref() == b"field" => {
-                let row = current_row.as_mut().ok_or_else(|| {
-                    DbOperationError::QueryFailed("MySQL XML field is outside a row".to_string())
-                })?;
-                if current_field.is_some() {
-                    return Err(DbOperationError::QueryFailed(
-                        "nested MySQL XML fields are not supported".to_string(),
-                    ));
-                }
-                let field = parse_mysql_field(&element)?;
-                row.push(field.finish());
-            }
-            Event::Text(text) => {
-                let decoded = text.decode().map_err(|error| {
-                    DbOperationError::QueryFailed(format!("invalid MySQL XML text: {error}"))
-                })?;
-                let text = unescape(&decoded).map_err(|error| {
-                    DbOperationError::QueryFailed(format!("invalid MySQL XML text: {error}"))
-                })?;
-                if let Some(field) = current_field.as_mut() {
-                    field.value.push_str(&text);
-                } else if !text.chars().all(char::is_whitespace) {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected text in MySQL XML result".to_string(),
-                    ));
-                }
-            }
-            Event::GeneralRef(reference) => {
-                let text = decode_mysql_xml_reference(&reference)?;
-                if let Some(field) = current_field.as_mut() {
-                    field.value.push_str(&text);
-                } else if !text.chars().all(char::is_whitespace) {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected text in MySQL XML result".to_string(),
-                    ));
-                }
-            }
-            Event::CData(data) => {
-                if let Some(field) = current_field.as_mut() {
-                    field
-                        .value
-                        .push_str(std::str::from_utf8(data.as_ref()).map_err(|error| {
-                            DbOperationError::QueryFailed(format!(
-                                "invalid MySQL XML text: {error}"
-                            ))
-                        })?);
-                } else {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected CDATA in MySQL XML result".to_string(),
-                    ));
-                }
-            }
-            Event::End(element) => match element.name().as_ref() {
-                b"field" => {
-                    let row = current_row.as_mut().ok_or_else(|| {
-                        DbOperationError::QueryFailed(
-                            "MySQL XML field is outside a row".to_string(),
-                        )
-                    })?;
-                    let field = current_field.take().ok_or_else(|| {
-                        DbOperationError::QueryFailed("unexpected MySQL XML field end".to_string())
-                    })?;
-                    row.push(field.finish());
-                }
-                b"row" => {
-                    let row = current_row.take().ok_or_else(|| {
-                        DbOperationError::QueryFailed("unexpected MySQL XML row end".to_string())
-                    })?;
-                    if columns.is_empty() {
-                        columns = row.iter().map(|(name, _)| name.clone()).collect();
-                    } else if row.len() != columns.len()
-                        || row
-                            .iter()
-                            .zip(&columns)
-                            .any(|((name, _), column)| name != column)
-                    {
-                        return Err(DbOperationError::QueryFailed(
-                            "MySQL XML rows have inconsistent fields".to_string(),
-                        ));
-                    }
-                    rows.push(row.into_iter().map(|(_, value)| value).collect());
-                }
-                b"resultset" => {
-                    if !in_resultset || current_row.is_some() || current_field.is_some() {
-                        return Err(DbOperationError::QueryFailed(
-                            "malformed MySQL XML resultset".to_string(),
-                        ));
-                    }
-                    in_resultset = false;
-                }
-                _ => {
-                    return Err(DbOperationError::QueryFailed(
-                        "unexpected MySQL XML closing element".to_string(),
-                    ));
-                }
-            },
-            Event::Decl(_) | Event::Comment(_) => {}
-            Event::Eof => break,
-            _ => {
-                return Err(DbOperationError::QueryFailed(
-                    "unexpected event in MySQL XML result".to_string(),
-                ));
-            }
-        }
-        buffer.clear();
-    }
-
-    if resultset_count != 1 || in_resultset || current_row.is_some() || current_field.is_some() {
-        return Err(DbOperationError::QueryFailed(
-            "MySQL XML result did not contain one complete resultset".to_string(),
-        ));
-    }
-    Ok(MysqlResultSet {
-        columns,
-        values: rows,
-    })
-}
-
-struct MysqlField {
-    name: String,
-    value: String,
-    is_null: bool,
-}
-
-impl MysqlField {
-    fn finish(self) -> (String, QueryValue) {
-        let value = if self.is_null {
-            QueryValue::Null
-        } else {
-            QueryValue::Text(self.value)
-        };
-        (self.name, value)
-    }
-
-    fn finish_raw(self) -> (String, String) {
-        let value = if self.is_null {
-            String::new()
-        } else {
-            self.value
-        };
-        (self.name, value)
-    }
-}
-
-fn parse_mysql_field(
-    element: &quick_xml::events::BytesStart<'_>,
-) -> Result<MysqlField, DbOperationError> {
-    let mut name = None;
-    let mut is_null = false;
-    for attribute in element.attributes() {
-        let attribute = attribute.map_err(|error| {
-            DbOperationError::QueryFailed(format!("invalid MySQL XML field: {error}"))
-        })?;
-        let value = attribute.unescape_value().map_err(|error| {
-            DbOperationError::QueryFailed(format!("invalid MySQL XML field: {error}"))
-        })?;
-        match attribute.key.as_ref() {
-            b"name" => name = Some(value.into_owned()),
-            b"xsi:nil" | b"nil" => is_null = matches!(value.as_ref(), "true" | "1"),
-            _ => {}
-        }
-    }
-    let name = name
-        .ok_or_else(|| DbOperationError::QueryFailed("MySQL XML field has no name".to_string()))?;
-    Ok(MysqlField {
-        name,
-        value: String::new(),
-        is_null,
-    })
-}
-
-#[cfg(test)]
-mod probe_tests {
-    use sabiql_app::model::connection::error::{ConnectionErrorInfo, ConnectionErrorKind};
-
-    use super::*;
-
-    #[test]
-    fn metadata_only_select_rejects_known_side_effects() {
-        for query in [
-            "SELECT value FROM items FOR UPDATE",
-            "SELECT GET_LOCK('sabiql', 0)",
-            "SELECT @value := 1",
-        ] {
-            assert!(
-                mysql_metadata_select_query(query, "__source", "__marker").is_err(),
-                "{query}"
-            );
-        }
-        assert!(mysql_metadata_select_query(
-            "WITH cte_rows AS (SELECT 1 AS first_alias) SELECT first_alias FROM cte_rows WHERE FALSE",
-            "__source",
-            "__marker"
-        )
-        .is_ok());
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT CONCAT('a', 'b') AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT CONCAT/**/('a', 'b') AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT sabiql_test.user_function() AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT sabiql_test/**/.user_function() AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT `user_function`/**/() AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT @session_value AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT INTERVAL(10, 1, 5) AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_err()
-        );
-        assert!(
-            mysql_metadata_select_query(
-            "WITH cte_rows(first_alias) AS (SELECT 1) SELECT first_alias FROM cte_rows WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_ok()
-        );
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT CASE (1) WHEN 1 THEN 'x' ELSE 'y' END AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_ok()
-        );
-        for query in [
-            "SELECT CAST(1 AS CHAR) AS value WHERE FALSE",
-            "SELECT CONVERT(1, CHAR) AS value WHERE FALSE",
-            "SELECT EXTRACT(YEAR FROM CURRENT_DATE) AS value WHERE FALSE",
-        ] {
-            assert!(
-                mysql_metadata_select_query(query, "__source", "__marker").is_err(),
-                "{query}"
-            );
-        }
-        assert!(
-            mysql_metadata_select_query(
-                "SELECT SLEEP(1) AS value WHERE FALSE",
-                "__source",
-                "__marker"
-            )
-            .is_ok()
-        );
-    }
-
-    #[test]
-    fn validates_supported_versions_and_sql_modes() {
-        assert!(is_oracle_mysql_cli_84_version("mysql  Ver 8.4.3 for macos"));
-        assert!(!is_oracle_mysql_cli_84_version(
-            "mysql  Ver 8.0.36 for macos"
-        ));
-        assert!(!is_oracle_mysql_cli_84_version(
-            "mysql  Ver 15.1 Distrib 10.11.8-MariaDB"
-        ));
-        assert!(!is_oracle_mysql_cli_84_version(
-            "mysql  Ver 8.4.3-3 for Linux (Percona Server)"
-        ));
-        assert!(is_oracle_mysql_server_84_version("8.4.3"));
-        assert!(!is_oracle_mysql_server_84_version("8.4.3-TiDB"));
-        assert!(!is_oracle_mysql_server_84_version("8.4.3-Percona"));
-        assert!(validate_sql_mode("STRICT_TRANS_TABLES").is_ok());
-        assert!(validate_sql_mode("STRICT_TRANS_TABLES,ANSI_QUOTES").is_err());
-    }
-
-    #[test]
-    fn uses_tcp_and_keeps_defaults_file_first() {
-        let args = mysql_probe_args(std::path::Path::new("/tmp/sabiql-mysql.cnf"));
-
-        assert_eq!(args[0], "--defaults-file=/tmp/sabiql-mysql.cnf");
-        assert_eq!(args[1], "--no-login-paths");
-        assert_eq!(args[2], "--protocol=TCP");
-        assert!(args.contains(&"--batch".to_string()));
-        assert!(args.contains(&"--raw".to_string()));
-        assert!(args.contains(&"--skip-column-names".to_string()));
-        assert!(args.contains(&"--binary-mode".to_string()));
-        assert!(args.contains(&"--skip-reconnect".to_string()));
-        assert!(
-            args.last()
-                .unwrap()
-                .starts_with("--execute=SELECT JSON_OBJECT")
-        );
-    }
-
-    #[test]
-    fn arguments_do_not_contain_credentials() {
-        let args = mysql_probe_args(std::path::Path::new("/tmp/sabiql-mysql.cnf"));
-
-        assert!(
-            args.iter()
-                .all(|argument| { !argument.contains("password") && !argument.contains("secret") })
-        );
-    }
-
-    #[test]
-    fn classifies_mysql_cli_timeout_errno_before_connection_refusal() {
-        assert!(matches!(
-            classify_mysql_probe_failure(
-                "ERROR 2003 (HY000): Can't connect to MySQL server on 'host:3306' (110)"
-                    .to_string()
-            ),
-            DbOperationError::Timeout(_)
-        ));
-        assert!(matches!(
-            classify_mysql_probe_failure(
-                "ERROR 2003 (HY000): Can't connect to MySQL server on 'host:3306' (111)"
-                    .to_string()
-            ),
-            DbOperationError::ConnectionFailed(_)
-        ));
-    }
-
-    #[test]
-    fn classifies_mysql_tls_probe_failure_for_connection_error() {
-        let error = classify_mysql_probe_failure(
-            "ERROR 2026 (HY000): SSL connection error: error:0A000086:SSL routines::certificate verify failed"
-                .to_string(),
-        );
-
-        assert_eq!(
-            ConnectionErrorInfo::from_db_operation_error(&error).kind,
-            ConnectionErrorKind::MySqlTlsHandshakeFailed
-        );
-    }
-}
-
-#[cfg(test)]
-mod query_tests {
-    use sabiql_app::model::connection::error::{ConnectionErrorInfo, ConnectionErrorKind};
-    use tempfile::tempdir;
-
-    use super::*;
-
-    #[test]
-    fn parses_mysql_xml_without_collapsing_value_boundaries() {
-        let xml = r#"<?xml version="1.0" encoding="utf-8"?>
-<resultset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-<row>
-  <field name="null" xsi:nil="true"/>
-  <field name="empty"></field>
-  <field name="text-null">NULL</field>
-  <field name="special">tab&#9;line&#10;slash\unicode 日本語</field>
-  <field name="json">{"a":[1,true]}</field>
-  <field name="binary-looking">0x41</field>
-</row>
-</resultset>"#;
-
-        let result = parse_mysql_xml(xml.as_bytes()).unwrap();
-
-        assert_eq!(
-            result.columns,
-            vec![
-                "null",
-                "empty",
-                "text-null",
-                "special",
-                "json",
-                "binary-looking"
-            ]
-        );
-        assert_eq!(
-            result.values,
-            vec![vec![
-                QueryValue::Null,
-                QueryValue::Text(String::new()),
-                QueryValue::Text("NULL".to_string()),
-                QueryValue::Text("tab\tline\nslash\\unicode 日本語".to_string()),
-                QueryValue::Text("{\"a\":[1,true]}".to_string()),
-                QueryValue::Text("0x41".to_string()),
-            ]]
-        );
-    }
-
-    #[test]
-    fn parses_numeric_and_binary_values_as_text() {
-        let xml = br#"<resultset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"><row>
-<field name="integer">18446744073709551615</field>
-<field name="decimal">12345678901234567890.123456789</field>
-<field name="float">1.25e+100</field>
-<field name="binary">0x00FF10</field>
-</row></resultset>"#;
-
-        let result = parse_mysql_xml(xml).unwrap();
-        assert!(
-            result.values[0]
-                .iter()
-                .all(|value| matches!(value, QueryValue::Text(_)))
-        );
-        assert_eq!(result.values[0][0].as_str(), Some("18446744073709551615"));
-        assert_eq!(result.values[0][3].as_str(), Some("0x00FF10"));
-    }
-
-    #[test]
-    fn rejects_multiple_resultsets_instead_of_guessing_the_last_one() {
-        let xml = br#"<resultset><row><field name="value">1</field></row></resultset>
-<resultset><row><field name="value">2</field></row></resultset>"#;
-
-        assert!(matches!(
-            parse_mysql_xml(xml),
-            Err(DbOperationError::QueryFailed(details))
-                if details.contains("more than one resultset")
-        ));
-    }
-
-    #[test]
-    fn accepts_xml_declaration_after_probe_separator_and_empty_resultsets() {
-        let xml = br#"
-<?xml version="1.0" encoding="utf-8"?>
-<resultset></resultset>
-"#;
-
-        let result = parse_mysql_xml(xml).unwrap();
-        assert!(result.columns.is_empty());
-        assert!(result.values.is_empty());
-    }
-
-    #[tokio::test]
-    async fn streams_mysql_xml_rows_into_csv_without_binary_type_inference() {
-        let xml = r#"<?xml version="1.0" encoding="utf-8"?>
-<resultset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"><row>
-<field name="comma">a,b</field>
-<field name="quote">a&quot;b</field>
-<field name="newline"><![CDATA[line1
-line2]]></field>
-<field name="tab">a	b</field>
-<field name="unicode">日本語</field>
-<field name="null" xsi:nil="true"/>
-<field name="empty"></field>
-<field name="binary">0x00FF</field>
-</row></resultset>"#;
-        let (mut input, output) = tokio::io::duplex(32);
-        let producer = tokio::spawn(async move {
-            input.write_all(xml.as_bytes()).await.unwrap();
-        });
-        let directory = tempdir().unwrap();
-        let path = directory.path().join("stream.csv");
-        let mut csv_writer = CsvFileWriter::create(path.clone()).await.unwrap();
-        let mut reader = Reader::from_reader(BufReader::new(output));
-        reader.config_mut().trim_text(false);
-
-        stream_mysql_xml_to_csv(&mut reader, &mut csv_writer)
-            .await
-            .unwrap();
-        csv_writer.finish().await.unwrap();
-        producer.await.unwrap();
-
-        assert_eq!(
-            fs::read_to_string(path).unwrap(),
-            "comma,quote,newline,tab,unicode,null,empty,binary\n\
-             \"a,b\",\"a\"\"b\",\"line1\n\
-             line2\",a\tb,日本語,,,0x00FF\n"
-                .replace("             ", "")
-        );
-    }
-
-    #[test]
-    fn csv_export_accepts_one_read_only_result_query() {
-        assert!(validate_mysql_export_query("SELECT 1", Some("app")).is_ok());
-        for query in ["TABLE users", "SHOW TABLES", "DESCRIBE users"] {
-            assert!(
-                validate_mysql_export_query(query, Some("app")).is_ok(),
-                "{query}"
-            );
-        }
-        assert!(matches!(
-            validate_mysql_export_query("INSERT INTO users VALUES (1)", Some("app")),
-            Err(DbOperationError::PermissionDenied(_))
-        ));
-        assert!(matches!(
-            validate_mysql_export_query("SELECT 1; SELECT 2", Some("app")),
-            Err(DbOperationError::UnsupportedOperation(details))
-                if details.contains("single read-only result")
-        ));
-    }
-
-    #[test]
-    fn mode_probe_requires_marker_and_allowed_mode_before_user_sql() {
-        let probe = MysqlResultSet {
-            columns: vec![
-                "__sabiql_probe".to_string(),
-                "__sabiql_sql_mode".to_string(),
-            ],
-            values: vec![vec![
-                QueryValue::Text("marker".to_string()),
-                QueryValue::Text("STRICT_TRANS_TABLES".to_string()),
-            ]],
-        };
-        assert!(validate_mode_probe(&probe, "marker").is_ok());
-
-        let mut unsupported = probe;
-        unsupported.values[0][1] = QueryValue::Text("ANSI_QUOTES".to_string());
-        assert!(matches!(
-            validate_mode_probe(&unsupported, "marker"),
-            Err(DbOperationError::UnsupportedOperation(details))
-                if details.contains(MYSQL_SQL_MODE_UNSUPPORTED_MARKER)
-        ));
-    }
-
-    #[test]
-    fn arguments_keep_credentials_out_of_argv() {
-        let args = mysql_query_args(std::path::Path::new("/tmp/sabiql-mysql.cnf"));
-
-        assert_eq!(args[0], "--defaults-file=/tmp/sabiql-mysql.cnf");
-        assert_eq!(args[1], "--no-login-paths");
-        for expected in [
-            "--xml",
-            "--binary-as-hex",
-            "--binary-mode",
-            "--unbuffered",
-            "--skip-reconnect",
-            "--default-character-set=utf8mb4",
-        ] {
-            assert!(args.contains(&expected.to_string()), "{expected}");
-        }
-        assert!(args.contains(&"--batch".to_string()));
-        assert!(args.iter().all(|argument| !argument.contains("password")));
-    }
-
-    #[test]
-    fn classifies_mysql_query_failures_by_server_error() {
-        assert!(matches!(
-            classify_mysql_query_failure(
-                b"ERROR 1045 (28000): Access denied for user 'app'@'localhost' (using password: YES)"
-            ),
-            DbOperationError::ConnectionFailed(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(b"ERROR 1142 (42000): command denied to user"),
-            DbOperationError::PermissionDenied(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(
-                b"ERROR 1044 (42000): Access denied for user 'app' to database 'app'"
-            ),
-            DbOperationError::PermissionDenied(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(b"ERROR 1146 (42S02): Table does not exist"),
-            DbOperationError::ObjectMissing(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(b"ERROR 1205 (HY000): Lock wait timeout exceeded"),
-            DbOperationError::LockTimeout(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(
-                b"ERROR 1452 (23000): Cannot add or update a child row: a foreign key constraint fails"
-            ),
-            DbOperationError::ForeignKeyViolation(_)
-        ));
-        let masked = classify_mysql_query_failure(b"ERROR password=secret");
-        assert!(!masked.masked_details().contains("secret"));
-    }
-
-    #[test]
-    fn classifies_mysql_tls_query_failures_as_connection_errors() {
-        let error = classify_mysql_query_failure(
-            b"ERROR 2026 (HY000): SSL connection error: error:0A000086:SSL routines::certificate verify failed",
-        );
-
-        assert_eq!(
-            ConnectionErrorInfo::from_db_operation_error(&error).kind,
-            ConnectionErrorKind::MySqlTlsHandshakeFailed
-        );
-        assert!(matches!(error, DbOperationError::ConnectionFailed(_)));
-    }
-}
-
-#[cfg(all(test, unix))]
-mod executor_tests {
-    use std::os::unix::fs::PermissionsExt;
-
-    use tempfile::TempDir;
-
-    use super::*;
-
-    async fn export_mysql_csv_with_program(
-        program: &OsStr,
-        option_file: &std::path::Path,
-        query: &str,
-        path: PathBuf,
-        execution_timeout: Duration,
-    ) -> Result<(), DbOperationError> {
-        let mut process = MysqlProcess::spawn_with_program(program, option_file)?;
-        let result = timeout(
-            execution_timeout,
-            run_mysql_export_process(&mut process, option_file, query, path),
-        )
-        .await;
-        match result {
-            Ok(Ok(())) => Ok(()),
-            Ok(Err(error)) => {
-                cleanup_mysql_process(&mut process).await;
-                Err(error)
-            }
-            Err(_) => {
-                cleanup_mysql_process(&mut process).await;
-                Err(DbOperationError::Timeout(
-                    "mysql query exceeded the execution timeout".to_string(),
-                ))
-            }
-        }
-    }
-
-    async fn run_mysql_adhoc_with_program(
-        program: &OsStr,
-        option_file: &std::path::Path,
-        query: &str,
-        access_mode: AccessMode,
-        execution_timeout: Duration,
-    ) -> Result<MysqlResultSet, DbOperationError> {
-        let mut process = MysqlProcess::spawn_with_program(program, option_file)?;
-        let result = timeout(
-            execution_timeout,
-            run_mysql_single_statement_process(&mut process, query, access_mode),
-        )
-        .await;
-        match result {
-            Ok(Ok(result_set)) => Ok(result_set),
-            Ok(Err(error)) => {
-                cleanup_mysql_process(&mut process).await;
-                Err(error)
-            }
-            Err(_) => {
-                cleanup_mysql_process(&mut process).await;
-                Err(DbOperationError::Timeout(
-                    "mysql query exceeded the execution timeout".to_string(),
-                ))
-            }
-        }
-    }
-
-    #[test]
-    fn failure_before_a_change_keeps_original_error() {
-        let error = query_failed_after_change(
-            DbOperationError::ForeignKeyViolation("foreign key failed".to_string()),
-            RefreshScope::None,
-        );
-
-        assert!(matches!(
-            error,
-            DbOperationError::ForeignKeyViolation(details) if details == "foreign key failed"
-        ));
-    }
-
-    fn fake_mysql(mode: &str) -> (TempDir, PathBuf, PathBuf) {
-        let directory = tempfile::tempdir().unwrap();
-        let option_file = directory.path().join("option.cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-        let program = directory.path().join("mysql");
-        let probe_response = match mode {
-            "missing" => "exit 0".to_string(),
-            "invalid" => {
-                "printf '%s\\n' '<resultset><row><field name=\"wrong\">x</field></row></resultset>'"
-                    .to_string()
-            }
-            "unsupported" => "printf '%s\\n' '<resultset><row><field name=\"__sabiql_probe\">'\"$marker\"'</field><field name=\"__sabiql_sql_mode\">ANSI_QUOTES</field></row></resultset>'".to_string(),
-            "timeout" => "while :; do :; done".to_string(),
-            _ => "printf '%s\\n' '<resultset><row><field name=\"__sabiql_probe\">'\"$marker\"'</field><field name=\"__sabiql_sql_mode\">STRICT_TRANS_TABLES</field></row></resultset>'".to_string(),
-        };
-        let user_response = if mode == "failure" {
-            "printf '%s\\n' '<resultset><row><field name=\"partial\">row</field></row></resultset>'\n    printf '%s\\n' 'ERROR 1064 (42000): syntax error' >&2\n    exit 1"
-        } else if mode == "no_result_failure" {
-            "printf '%s\\n' 'ERROR 1054 (42S22): Unknown column missing_column' >&2\n    exit 1"
-        } else {
-            "printf '%s\\n' '<resultset><row><field name=\"value\">ok</field></row></resultset>'"
-        };
-        let session_failure = if mode == "read_only_failure" {
-            "printf '%s\\n' 'ERROR 1227 (42000): access denied to set transaction read only' >&2\n      exit 1"
-        } else {
-            ""
-        };
-        let script = format!(
-            r#"#!/bin/sh
-option=$(printf '%s\n' "$1" | sed 's/^--defaults-file=//')
-log="$option.log"
-phase=probe
-while IFS= read -r line; do
-  printf '%s\n' "$line" >> "$log"
-  [ "$line" = ";" ] && continue
-  if [ "$phase" = probe ]; then
-    marker=$(printf '%s\n' "$line" | sed "s/.*SELECT '\\([^']*\\)'.*/\\1/")
-    {probe_response}
-    phase=user
-  else
-    case "$line" in
-      "SET SESSION TRANSACTION READ ONLY")
-        {session_failure}
-        ;;
-      *__sabiql_session_marker*)
-        marker=$(printf '%s\n' "$line" | sed "s/.*SELECT '\\([^']*\\)' AS __sabiql_session_marker.*/\\1/")
-        printf '%s\n' '<resultset><row><field name="__sabiql_session_marker">'"$marker"'</field></row></resultset>'
-        ;;
-      *)
-        {user_response}
-        exit 0
-        ;;
-    esac
-  fi
-done
-"#,
-        );
-        fs::write(&program, script).unwrap();
-        let mut permissions = fs::metadata(&program).unwrap().permissions();
-        permissions.set_mode(0o755);
-        fs::set_permissions(&program, permissions).unwrap();
-        (directory, program, log_file)
-    }
-
-    fn fake_mysql_multi() -> (TempDir, PathBuf, PathBuf) {
-        fake_mysql_multi_with_mode(false, None)
-    }
-
-    fn fake_mysql_multi_with_marker_failure() -> (TempDir, PathBuf, PathBuf) {
-        fake_mysql_multi_with_mode(true, None)
-    }
-
-    fn fake_mysql_multi_with_statement_failure(error: &str) -> (TempDir, PathBuf, PathBuf) {
-        fake_mysql_multi_with_mode(false, Some(error))
-    }
-
-    fn fake_mysql_multi_with_mode(
-        marker_failure: bool,
-        statement_error: Option<&str>,
-    ) -> (TempDir, PathBuf, PathBuf) {
-        let directory = tempfile::tempdir().unwrap();
-        let option_file = directory.path().join("option.cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let program = directory.path().join("mysql");
-        let update_response = statement_error.map_or_else(
-            || {
-                "printf '%s\\n' '<resultset><row><field name=\"affected\">ok</field></row></resultset>'"
-                    .to_string()
-            },
-            |error| format!("printf '%s\\n' '{error}' >&2"),
-        );
-        let marker_response = if marker_failure {
-            "printf '%s\\n' '<resultset><row><field name=\"wrong\">x</field></row></resultset>'"
-        } else {
-            "marker=$(printf '%s\\n' \"$line\" | sed \"s/.*SELECT '\\\\([^']*\\\\)' AS __sabiql_marker.*/\\\\1/\")
-      rows=0
-      case \"$line\" in *ROW_COUNT\\(\\)* ) rows=3 ;; esac
-      printf '%s\\n' '<resultset><row><field name=\"__sabiql_marker\">'\"$marker\"'</field><field name=\"affected_rows\">'\"$rows\"'</field></row></resultset>'
-      if [ \"$pending_error\" = 1 ]; then
-        sleep 0.05
-        printf '%s\\n' 'ERROR 1054 (42S22): Unknown column missing_column' >&2
-        pending_error=0
-      fi"
-        };
-        let script = format!(
-            r#"#!/bin/sh
-option=$(printf '%s\n' "$1" | sed 's/^--defaults-file=//')
-log="$option.log"
-printf 'process=%s\n' "$$" >> "$log"
-pending_error=0
-while IFS= read -r line; do
-  printf '%s\n' "$line" >> "$log"
-  case "$line" in
-    *__sabiql_probe*)
-      marker=$(printf '%s\n' "$line" | sed "s/.*SELECT '\\([^']*\\)' AS __sabiql_probe.*/\\1/")
-      printf '%s\n' '<resultset><row><field name="__sabiql_probe">'"$marker"'</field><field name="__sabiql_sql_mode">STRICT_TRANS_TABLES</field></row></resultset>'
-      ;;
-    "SET SESSION TRANSACTION READ ONLY")
-      ;;
-    *__sabiql_session_marker*)
-      marker=$(printf '%s\n' "$line" | sed "s/.*SELECT '\\([^']*\\)' AS __sabiql_session_marker.*/\\1/")
-      printf '%s\n' '<resultset><row><field name="__sabiql_session_marker">'"$marker"'</field></row></resultset>'
-      ;;
-      *__sabiql_marker*)
-      {marker_response}
-      ;;
-    *missing_column*)
-      pending_error=1
-      ;;
-    *SELECT*)
-      value=one
-      case "$line" in *SELECT\ 2*) value=two ;; esac
-      printf '%s\n' '<resultset><row><field name="value">'"$value"'</field></row></resultset>'
-      ;;
-    *UPDATE*)
-      {update_response}
-      ;;
-    *)
-      printf '%s\n' '<resultset></resultset>'
-      ;;
-  esac
-done
-"#,
-        );
-        fs::write(&program, script).unwrap();
-        let mut permissions = fs::metadata(&program).unwrap().permissions();
-        permissions.set_mode(0o755);
-        fs::set_permissions(&program, permissions).unwrap();
-        (directory, program, option_file)
-    }
-
-    #[tokio::test]
-    async fn sends_user_sql_only_after_a_valid_mode_probe() {
-        let (_directory, program, log_file) = fake_mysql("success");
-        let option_file = log_file.with_extension("cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let result = run_mysql_adhoc_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 123",
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(result.columns, vec!["value"]);
-        assert_eq!(result.values[0][0].as_str(), Some("ok"));
-        let log = fs::read_to_string(format!("{}.log", option_file.display())).unwrap();
-        assert!(log.contains("__sabiql_probe"));
-        assert!(log.contains("SELECT 123"));
-        assert!(!log.contains(MYSQL_READ_ONLY_STATEMENT));
-    }
-
-    #[tokio::test]
-    async fn configures_read_only_session_before_user_sql() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-        let statements = split_mysql_statements("SELECT 2")
-            .unwrap()
-            .into_iter()
-            .map(|sql| classify_mysql_statement(&sql).unwrap())
-            .collect::<Vec<_>>();
-
-        let result = run_mysql_adhoc_with_program_and_statements(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 2",
-            &statements,
-            AccessMode::ReadOnly,
-            Duration::from_secs(5),
-        )
-        .await
-        .unwrap_or_else(|error| {
-            let log = fs::read_to_string(&log_file).unwrap_or_default();
-            panic!("read-only execution failed: {error:?}; log: {log}");
-        });
-
-        assert_eq!(
-            result.result_set.unwrap().values[0][0].as_str(),
-            Some("two")
-        );
-        let log = fs::read_to_string(log_file).unwrap();
-        let session_index = log
-            .find(MYSQL_READ_ONLY_STATEMENT)
-            .expect("read-only session statement");
-        let user_index = log.find("SELECT 2").expect("user statement");
-        assert!(session_index < user_index, "{log}");
-        assert!(log.contains(MYSQL_SESSION_MARKER_COLUMN));
-    }
-
-    #[tokio::test]
-    async fn metadata_session_reuses_one_process_for_ordered_resultsets() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let mut session =
-            MysqlMetadataSession::spawn_with_program(OsStr::new(&program), &option_file)
-                .expect("spawn fake mysql");
-
-        session.probe().await.expect("mode probe");
-        for query in [
-            "SELECT TABLES",
-            "SELECT COLUMNS",
-            "SELECT INDEXES",
-            "SELECT FOREIGN_KEYS",
-            "SELECT TRIGGERS",
-            "SHOW CREATE TABLE items",
-        ] {
-            session.execute(query).await.expect("metadata resultset");
-        }
-        session.finish().await.expect("finish fake mysql");
-
-        let log = fs::read_to_string(format!("{}.log", option_file.display())).unwrap();
-        assert_eq!(
-            log.lines()
-                .filter(|line| line.starts_with("process="))
-                .count(),
-            1
-        );
-        let positions = [
-            "__sabiql_probe",
-            "SELECT TABLES",
-            "SELECT COLUMNS",
-            "SELECT INDEXES",
-            "SELECT FOREIGN_KEYS",
-            "SELECT TRIGGERS",
-            "SHOW CREATE TABLE items",
-        ]
-        .into_iter()
-        .map(|query| log.find(query).expect("query in transcript"))
-        .collect::<Vec<_>>();
-        assert!(positions.windows(2).all(|pair| pair[0] < pair[1]), "{log}");
-    }
-
-    #[tokio::test]
-    async fn read_only_session_failure_never_writes_user_sql() {
-        let (_directory, program, log_file) = fake_mysql("read_only_failure");
-        let option_file = log_file.with_extension("cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let result = run_mysql_adhoc_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 123",
-            AccessMode::ReadOnly,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        assert!(result.is_err());
-        let log = fs::read_to_string(format!("{}.log", option_file.display())).unwrap();
-        assert!(log.contains(MYSQL_READ_ONLY_STATEMENT));
-        assert!(!log.contains("SELECT 123"), "{log}");
-    }
-
-    #[tokio::test]
-    async fn generated_preview_and_metadata_queries_skip_read_only_session_setup() {
-        for query in [
-            "SELECT id FROM app.items ORDER BY id LIMIT 10 OFFSET 0",
-            "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES",
-        ] {
-            let (_directory, program, option_file) = fake_mysql_multi();
-            let statements = split_mysql_statements(query)
-                .unwrap()
-                .into_iter()
-                .map(|sql| classify_mysql_statement(&sql).unwrap())
-                .collect::<Vec<_>>();
-
-            run_mysql_adhoc_with_program_and_statements(
-                OsStr::new(&program),
-                &option_file,
-                query,
-                &statements,
-                AccessMode::ReadWrite,
-                Duration::from_secs(5),
-            )
-            .await
-            .unwrap();
-
-            let log = fs::read_to_string(format!("{}.log", option_file.display())).unwrap();
-            assert!(!log.contains(MYSQL_READ_ONLY_STATEMENT), "{query}: {log}");
-        }
-    }
-
-    #[test]
-    fn read_only_rejects_temporary_table_dml_before_starting_mysql() {
-        let (_directory, _program, option_file) = fake_mysql_multi();
-        let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-        let query = "CREATE TEMPORARY TABLE temp_items (id INT); INSERT INTO temp_items VALUES (1); DROP TEMPORARY TABLE temp_items";
-
-        let result = validate_mysql_multi_query(query, Some("app"), AccessMode::ReadOnly);
-
-        assert!(matches!(
-            result,
-            Err(DbOperationError::PermissionDenied(details))
-                if details.contains("read-only mode blocks MySQL write statements")
-        ));
-        assert!(!log_file.exists());
-    }
-
-    #[test]
-    fn read_only_rejects_read_write_overrides_before_starting_mysql() {
-        for query in [
-            "SET SESSION TRANSACTION READ WRITE",
-            "START TRANSACTION READ WRITE",
-        ] {
-            let (_directory, _program, option_file) = fake_mysql_multi();
-            let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-
-            let result = validate_mysql_multi_query(query, Some("app"), AccessMode::ReadOnly);
-
-            assert!(matches!(
-                result,
-                Err(DbOperationError::UnsupportedOperation(_))
-            ));
-            assert!(!log_file.exists(), "{query}");
-        }
-    }
-
-    #[test]
-    fn rejects_empty_metadata_fallback_after_temporary_table_creation() {
-        for query in [
-            "CREATE TEMPORARY TABLE temp_items (id INT); DESCRIBE temp_items 'missing'; DROP TEMPORARY TABLE temp_items",
-            "CREATE TEMPORARY TABLE temp_items (id INT); SHOW COLUMNS FROM temp_items LIKE 'missing'; DROP TEMPORARY TABLE temp_items",
-        ] {
-            let statements = validate_mysql_multi_query(query, Some("app"), AccessMode::ReadWrite)
-                .expect("query should be classified before the session-state check");
-
-            assert!(mysql_metadata_fallback_has_unsupported_session_state(
-                &statements
-            ));
-        }
-
-        let statements = validate_mysql_multi_query(
-            "SHOW COLUMNS FROM items",
-            Some("app"),
-            AccessMode::ReadWrite,
-        )
-        .expect("single SHOW should be classified");
-        assert!(!mysql_metadata_fallback_has_unsupported_session_state(
-            &statements
-        ));
-    }
-
-    #[tokio::test]
-    async fn exports_mysql_xml_rows_through_the_shared_csv_writer() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let path = option_file.with_file_name("export.csv");
-
-        export_mysql_csv_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 1",
-            path.clone(),
-            Duration::from_secs(5),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(fs::read_to_string(path).unwrap(), "value\none\n");
-    }
-
-    #[tokio::test]
-    async fn export_configures_read_only_session_before_user_sql() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-        let path = option_file.with_file_name("export.csv");
-
-        export_mysql_csv_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 1",
-            path,
-            Duration::from_secs(5),
-        )
-        .await
-        .unwrap();
-
-        let log = fs::read_to_string(log_file).unwrap();
-        let session_index = log
-            .find(MYSQL_READ_ONLY_STATEMENT)
-            .expect("read-only session statement");
-        let user_index = log.find("SELECT 1").expect("user statement");
-        assert!(session_index < user_index, "{log}");
-        assert!(log.contains(MYSQL_SESSION_MARKER_COLUMN));
-    }
-
-    #[tokio::test]
-    async fn export_read_only_session_failure_never_writes_user_sql_or_partial_file() {
-        let (_directory, program, option_file) = fake_mysql("read_only_failure");
-        let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-        let output_directory = tempfile::tempdir().unwrap();
-        let final_path = output_directory.path().join("export.csv");
-
-        let result = export_to_path(final_path.clone(), |path| {
-            export_mysql_csv_with_program(
-                OsStr::new(&program),
-                &option_file,
-                "SELECT 123",
-                path,
-                Duration::from_secs(5),
-            )
-        })
-        .await;
-
-        assert!(result.is_err());
-        let log = fs::read_to_string(log_file).unwrap();
-        assert!(log.contains(MYSQL_READ_ONLY_STATEMENT));
-        assert!(!log.contains("SELECT 123"), "{log}");
-        assert!(!final_path.exists());
-        assert_eq!(output_directory.path().read_dir().unwrap().count(), 0);
-    }
-
-    #[tokio::test]
-    async fn export_failure_removes_the_partial_file() {
-        let (_directory, program, option_file) = fake_mysql("failure");
-        let output_directory = tempfile::tempdir().unwrap();
-        let final_path = output_directory.path().join("export.csv");
-
-        let result = export_to_path(final_path.clone(), |path| {
-            export_mysql_csv_with_program(
-                OsStr::new(&program),
-                &option_file,
-                "SELECT 1",
-                path,
-                Duration::from_secs(5),
-            )
-        })
-        .await;
-
-        assert!(result.is_err());
-        assert!(!final_path.exists());
-        assert_eq!(output_directory.path().read_dir().unwrap().count(), 0);
-    }
-
-    #[tokio::test]
-    async fn export_timeout_kills_the_process_and_removes_the_partial_file() {
-        let (_directory, program, option_file) = fake_mysql("timeout");
-        let output_directory = tempfile::tempdir().unwrap();
-        let final_path = output_directory.path().join("export.csv");
-
-        let result = export_to_path(final_path.clone(), |path| {
-            export_mysql_csv_with_program(
-                OsStr::new(&program),
-                &option_file,
-                "SELECT 1",
-                path,
-                Duration::from_millis(50),
-            )
-        })
-        .await;
-
-        assert!(matches!(result, Err(DbOperationError::Timeout(_))));
-        assert!(!final_path.exists());
-        assert_eq!(output_directory.path().read_dir().unwrap().count(), 0);
-    }
-
-    #[test]
-    fn frames_one_xml_resultset_and_preserves_following_output() {
-        let mut buffer = b"    -> <?xml version=\"1.0\"?>\n<resultset></resultset>\r\n    -> <?xml version=\"1.0\"?>\n<resultset>"
-            .to_vec();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-
-        assert_eq!(
-            scanner.take(&mut buffer),
-            Some(b"<resultset></resultset>".to_vec())
-        );
-        assert_eq!(
-            scanner.take(&mut buffer),
-            None,
-            "an incomplete following frame must remain buffered"
-        );
-        assert!(buffer.starts_with(b"\r\n    -> <?xml"));
-    }
-
-    #[test]
-    fn frames_resultset_after_mysql_cli_text() {
-        let mut buffer =
-            b"SELECT 1;\n<?xml version=\"1.0\"?>\nquery text\n<resultset></resultset>".to_vec();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-
-        assert_eq!(
-            scanner.take(&mut buffer),
-            Some(b"<resultset></resultset>".to_vec())
-        );
-        assert!(buffer.is_empty());
-    }
-
-    #[tokio::test]
-    async fn probe_failure_never_writes_user_sql() {
-        for mode in ["unsupported", "invalid", "missing"] {
-            let (_directory, program, log_file) = fake_mysql(mode);
-            let option_file = log_file.with_extension("cnf");
-            fs::write(&option_file, "[client]\n").unwrap();
-            let result = run_mysql_adhoc_with_program(
-                OsStr::new(&program),
-                &option_file,
-                "SELECT 123",
-                AccessMode::ReadWrite,
-                Duration::from_secs(5),
-            )
-            .await;
-            assert!(result.is_err(), "{mode}");
-            let log =
-                fs::read_to_string(format!("{}.log", option_file.display())).unwrap_or_default();
-            assert!(!log.contains("SELECT 123"), "{mode}: {log}");
-        }
-    }
-
-    #[tokio::test]
-    async fn probe_timeout_kills_the_process_and_discards_output() {
-        let (_directory, program, log_file) = fake_mysql("timeout");
-        let option_file = log_file.with_extension("cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let result = run_mysql_adhoc_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 123",
-            AccessMode::ReadWrite,
-            Duration::from_millis(50),
-        )
-        .await;
-
-        assert!(matches!(result, Err(DbOperationError::Timeout(_))));
-        let log = fs::read_to_string(format!("{}.log", option_file.display())).unwrap_or_default();
-        assert!(!log.contains("SELECT 123"));
-    }
-
-    #[tokio::test]
-    async fn nonzero_cli_exit_discards_any_collected_stdout() {
-        let (_directory, program, log_file) = fake_mysql("failure");
-        let option_file = log_file.with_extension("cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let result = run_mysql_adhoc_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 123",
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        assert!(matches!(result, Err(DbOperationError::QueryFailed(_))));
-    }
-
-    #[tokio::test]
-    async fn classifies_cli_error_when_no_resultset_is_emitted() {
-        let (_directory, program, log_file) = fake_mysql("no_result_failure");
-        let option_file = log_file.with_extension("cnf");
-        fs::write(&option_file, "[client]\n").unwrap();
-        let result = run_mysql_adhoc_with_program(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT 123",
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        assert!(matches!(
-            result,
-            Err(DbOperationError::QueryFailed(details))
-                if details.contains("missing_column")
-        ));
-    }
-
-    #[tokio::test]
-    async fn executes_each_statement_and_returns_the_last_user_result() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let log_file = PathBuf::from(format!("{}.log", option_file.display()));
-        let statements = split_mysql_statements("UPDATE items SET value = 1; SELECT 2")
-            .unwrap()
-            .into_iter()
-            .map(|sql| classify_mysql_statement(&sql).unwrap())
-            .collect::<Vec<_>>();
-
-        let result = run_mysql_adhoc_with_program_and_statements(
-            OsStr::new(&program),
-            &option_file,
-            "UPDATE items SET value = 1; SELECT 2",
-            &statements,
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await
-        .unwrap_or_else(|error| panic!("multi execution failed: {error:?}"));
-
-        assert_eq!(
-            result.result_set,
-            Some(MysqlResultSet {
-                columns: vec!["value".to_string()],
-                values: vec![vec![QueryValue::Text("two".to_string())]],
-            })
-        );
-        assert_eq!(result.command_tag, Some(CommandTag::Update(3)));
-        assert_eq!(result.refresh_scope, RefreshScope::Data);
-        let log = fs::read_to_string(log_file).unwrap();
-        assert!(log.contains("UPDATE items SET value = 1"));
-        assert!(log.matches("__sabiql_marker").count() >= 2);
-    }
-
-    #[tokio::test]
-    async fn marker_failure_after_a_change_refreshes_the_current_scope() {
-        let (_directory, program, option_file) = fake_mysql_multi_with_marker_failure();
-        let statements = split_mysql_statements("UPDATE items SET value = 1")
-            .unwrap()
-            .into_iter()
-            .map(|sql| classify_mysql_statement(&sql).unwrap())
-            .collect::<Vec<_>>();
-
-        let result = run_mysql_adhoc_with_program_and_statements(
-            OsStr::new(&program),
-            &option_file,
-            "UPDATE items SET value = 1",
-            &statements,
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        assert!(matches!(
-            result,
-            Err(DbOperationError::QueryFailedAfterChange {
-                source,
-                refresh_scope: RefreshScope::Data,
-                ..
-            }) if matches!(&*source, DbOperationError::QueryFailed(_))
-        ));
-    }
-
-    #[tokio::test]
-    async fn first_change_statement_failure_keeps_the_classified_error_unwrapped() {
-        for (details, summary) in [
-            (
-                "ERROR 1142 (42000): command denied to user",
-                "Permission denied",
-            ),
-            (
-                "ERROR 1062 (23000): Duplicate entry duplicate_value for key PRIMARY",
-                "Unique constraint violation",
-            ),
-            (
-                "ERROR 1452 (23000): Cannot add or update a child row: a foreign key constraint fails",
-                "Foreign key constraint violation",
-            ),
-            (
-                "ERROR 1205 (HY000): Lock wait timeout exceeded",
-                "Operation blocked by lock or timeout",
-            ),
-        ] {
-            let (_directory, program, option_file) =
-                fake_mysql_multi_with_statement_failure(details);
-            let statements = split_mysql_statements("UPDATE items SET value = 1")
-                .unwrap()
-                .into_iter()
-                .map(|sql| classify_mysql_statement(&sql).unwrap())
-                .collect::<Vec<_>>();
-
-            let result = run_mysql_adhoc_with_program_and_statements(
-                OsStr::new(&program),
-                &option_file,
-                "UPDATE items SET value = 1",
-                &statements,
-                AccessMode::ReadWrite,
-                Duration::from_secs(5),
-            )
-            .await;
-            let Err(error) = result else {
-                panic!("expected the fake MySQL statement to fail");
-            };
-
-            assert_eq!(error.summary(), summary);
-            assert!(!matches!(
-                error,
-                DbOperationError::QueryFailedAfterChange { .. }
-            ));
-        }
-    }
-
-    #[tokio::test]
-    async fn marks_a_later_failure_after_a_confirmed_change_for_refresh() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let statements =
-            split_mysql_statements("UPDATE items SET value = 1; SELECT missing_column FROM items")
-                .unwrap()
-                .into_iter()
-                .map(|sql| classify_mysql_statement(&sql).unwrap())
-                .collect::<Vec<_>>();
-
-        let result = run_mysql_adhoc_with_program_and_statements(
-            OsStr::new(&program),
-            &option_file,
-            "UPDATE items SET value = 1; SELECT missing_column FROM items",
-            &statements,
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        assert!(matches!(
-            result,
-            Err(DbOperationError::QueryFailedAfterChange {
-                source,
-                refresh_scope: RefreshScope::Data,
-                ..
-            }) if matches!(&*source, DbOperationError::QueryFailed(_))
-        ));
-    }
-
-    #[tokio::test]
-    async fn rejects_error_reported_after_row_count_marker() {
-        let (_directory, program, option_file) = fake_mysql_multi();
-        let statements = split_mysql_statements("SELECT missing_column FROM items")
-            .unwrap()
-            .into_iter()
-            .map(|sql| classify_mysql_statement(&sql).unwrap())
-            .collect::<Vec<_>>();
-
-        let result = run_mysql_adhoc_with_program_and_statements(
-            OsStr::new(&program),
-            &option_file,
-            "SELECT missing_column FROM items",
-            &statements,
-            AccessMode::ReadWrite,
-            Duration::from_secs(5),
-        )
-        .await;
-
-        assert!(matches!(
-            result,
-            Err(DbOperationError::QueryFailed(details))
-                if details.contains("missing_column")
-        ));
-    }
-
-    #[test]
-    fn transaction_rollback_removes_pending_data_tag() {
-        let events = vec![
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Begin,
-                target: None,
-                tag: CommandTag::Begin,
-            },
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Update { has_where: true },
-                target: Some("items".to_string()),
-                tag: CommandTag::Update(1),
-            },
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Rollback,
-                target: None,
-                tag: CommandTag::Rollback,
-            },
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Select,
-                target: None,
-                tag: CommandTag::Select(1),
-            },
-        ];
-
-        assert_eq!(
-            aggregate_mysql_command_tag(&events),
-            Some(CommandTag::Select(1))
-        );
-    }
-
-    #[test]
-    fn ddl_implicit_commit_keeps_prior_data_change() {
-        let events = vec![
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Begin,
-                target: None,
-                tag: CommandTag::Begin,
-            },
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Insert,
-                target: Some("items".to_string()),
-                tag: CommandTag::Insert(1),
-            },
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::CreateTable { temporary: false },
-                target: Some("created".to_string()),
-                tag: CommandTag::Create("TABLE".to_string()),
-            },
-            MysqlCommandEvent {
-                kind: MysqlStatementKind::Rollback,
-                target: None,
-                tag: CommandTag::Rollback,
-            },
-        ];
-
-        assert_eq!(
-            aggregate_mysql_command_tag(&events),
-            Some(CommandTag::Create("TABLE".to_string()))
-        );
-    }
-}
-
-#[cfg(test)]
-mod resultset_frame_tests {
-    use super::*;
-
-    #[test]
-    fn extracts_one_frame_when_end_delimiter_crosses_4k_chunk_boundary() {
-        let delimiter_start = 4096 - 3;
-        let mut expected = MYSQL_RESULTSET_START.to_vec();
-        expected.resize(delimiter_start, b'x');
-        expected.extend_from_slice(MYSQL_RESULTSET_END);
-
-        let mut buffer = Vec::new();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-        let mut frames = Vec::new();
-        for chunk in expected.chunks(4096) {
-            buffer.extend_from_slice(chunk);
-            if let Some(frame) = scanner.take(&mut buffer) {
-                frames.push(frame);
-            }
-        }
-
-        assert_eq!(frames, vec![expected]);
-        assert!(buffer.is_empty());
-        assert_eq!(scanner.take(&mut buffer), None);
-    }
-
-    #[test]
-    fn extracts_large_resultset_from_small_chunks() {
-        let mut expected = MYSQL_RESULTSET_START.to_vec();
-        expected.extend_from_slice(b"<row><field name=\"value\">");
-        expected.extend(vec![b'x'; 128 * 1024]);
-        expected.extend_from_slice(b"</field></row>");
-        expected.extend_from_slice(MYSQL_RESULTSET_END);
-
-        let mut buffer = Vec::new();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-        let mut frames = Vec::new();
-        for chunk in expected.chunks(37) {
-            buffer.extend_from_slice(chunk);
-            if let Some(frame) = scanner.take(&mut buffer) {
-                frames.push(frame);
-            }
-        }
-
-        assert_eq!(frames, vec![expected]);
-        assert!(buffer.is_empty());
-    }
-
-    #[test]
-    fn drains_frames_in_order_without_skipping_the_following_frame() {
-        let first = b"<resultset><row><field name=\"value\">one</field></row></resultset>";
-        let second = b"noise<resultset><row><field name=\"value\">two</field></row></resultset>";
-        let mut input = first.to_vec();
-        input.extend_from_slice(second);
-        let mut buffer = Vec::new();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-        let mut frames = Vec::new();
-
-        for chunk in input.chunks(11) {
-            buffer.extend_from_slice(chunk);
-            while let Some(frame) = scanner.take(&mut buffer) {
-                frames.push(frame);
-            }
-        }
-
-        assert_eq!(frames, vec![first.to_vec(), second[5..].to_vec()]);
-        assert!(buffer.is_empty());
-    }
-
-    #[test]
-    fn delimiter_prefix_in_field_text_does_not_end_the_frame() {
-        let expected = b"<resultset><row><field name=\"value\">literal </resultset prefix</field></row></resultset>";
-        let mut buffer = expected.to_vec();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-
-        assert_eq!(scanner.take(&mut buffer), Some(expected.to_vec()));
-        assert!(buffer.is_empty());
-    }
-
-    #[test]
-    fn resultset_field_error_text_is_not_classified_as_cli_error() {
-        let mut buffer = br#"<resultset><row><field name="message">line 1
-ERROR 1146 (42S02): this is a cell value</field></row></resultset>"#
-            .to_vec();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-
-        let frame = take_mysql_pty_resultset_frame(&mut buffer, &mut scanner).unwrap();
-
-        assert!(frame.is_some());
-        assert!(buffer.is_empty());
-    }
-
-    #[test]
-    fn cli_error_before_resultset_frame_is_still_rejected() {
-        let mut buffer =
-            b"ERROR 1054 (42S22): Unknown column\n<resultset><row></row></resultset>".to_vec();
-        let mut scanner = MysqlResultsetFrameScanner::default();
-
-        let result = take_mysql_pty_resultset_frame(&mut buffer, &mut scanner);
-
-        assert!(matches!(result, Err(DbOperationError::QueryFailed(_))));
-        assert_eq!(
-            buffer,
-            b"ERROR 1054 (42S22): Unknown column\n<resultset><row></row></resultset>"
-        );
-    }
-
-    #[test]
-    fn error_before_resultset_frame_is_not_accepted() {
-        let mut buffer = b"<resultset><row></row></resultset>".to_vec();
-        let error = b"ERROR 1054 (42S22): Unknown column missing_column\n";
-        let mut scanner = MysqlResultsetFrameScanner::default();
-
-        assert!(matches!(
-            take_mysql_resultset_frame_after_error_check(&mut buffer, error, &mut scanner),
-            Err(DbOperationError::QueryFailed(_))
-        ));
-        assert_eq!(buffer, b"<resultset><row></row></resultset>");
-    }
-}
-
-#[cfg(all(test, not(unix)))]
-mod pipe_executor_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn pipe_errors_are_checked_before_resultset_frames() {
-        let (mut stdout_writer, mut stdout_reader) = tokio::io::duplex(1024);
-        let (mut stderr_writer, mut stderr_reader) = tokio::io::duplex(1024);
-        stdout_writer
-            .write_all(b"<resultset><row></row></resultset>")
-            .await
-            .unwrap();
-        stderr_writer
-            .write_all(b"ERROR 1054 (42S22): Unknown column missing_column\n")
-            .await
-            .unwrap();
-        drop(stdout_writer);
-        drop(stderr_writer);
-        let mut frame_scanner = MysqlResultsetFrameScanner::default();
-
-        let result = read_one_mysql_resultset_from_pipes(
-            &mut stdout_reader,
-            &mut stderr_reader,
-            &mut Vec::new(),
-            &mut Vec::new(),
-            &mut frame_scanner,
-        )
-        .await;
-
-        assert!(matches!(result, Err(DbOperationError::QueryFailed(_))));
     }
 }
