+++ src/infra/adapters/mysql/cli/probe.rs
@@ -1,0 +1,325 @@
+#[derive(Debug, Deserialize)]
+pub(super) struct MySqlProbeResponse {
+    pub(super) database: Option<String>,
+    pub(super) user: String,
+    pub(super) version: String,
+    pub(super) sql_mode: String,
+}
+
+fn contains_unsupported_mysql_product(value: &str) -> bool {
+    let lower = value.to_ascii_lowercase();
+    ["mariadb", "percona", "tidb", "vitess", "aurora"]
+        .iter()
+        .any(|product| lower.contains(product))
+}
+
+pub(super) fn is_oracle_mysql_cli_84_version(value: &str) -> bool {
+    let lower = value.to_ascii_lowercase();
+    lower.contains("mysql")
+        && !contains_unsupported_mysql_product(value)
+        && version_major_minor(value) == Some((8, 4))
+}
+
+pub(super) fn is_oracle_mysql_server_84_version(value: &str) -> bool {
+    !contains_unsupported_mysql_product(value) && version_major_minor(value) == Some((8, 4))
+}
+
+pub(super) fn validate_server_version(version: &str) -> Result<(), DbOperationError> {
+    if is_oracle_mysql_server_84_version(version) {
+        Ok(())
+    } else {
+        Err(DbOperationError::UnsupportedOperation(format!(
+            "{MYSQL_SERVER_VERSION_REQUIRED_MARKER}: {version}"
+        )))
+    }
+}
+
+pub(super) fn validate_sql_mode(sql_mode: &str) -> Result<(), DbOperationError> {
+    let unsupported = sql_mode.split(',').map(str::trim).any(|mode| {
+        mode.eq_ignore_ascii_case("NO_BACKSLASH_ESCAPES")
+            || mode.eq_ignore_ascii_case("ANSI_QUOTES")
+    });
+    if unsupported {
+        Err(DbOperationError::UnsupportedOperation(format!(
+            "{MYSQL_SQL_MODE_UNSUPPORTED_MARKER}: {sql_mode}"
+        )))
+    } else {
+        Ok(())
+    }
+}
+
+fn version_major_minor(value: &str) -> Option<(u32, u32)> {
+    let mut numbers = value
+        .split(|character: char| !character.is_ascii_digit())
+        .filter(|part| !part.is_empty())
+        .filter_map(|part| part.parse::<u32>().ok());
+    Some((numbers.next()?, numbers.next()?))
+}
+
+pub(super) fn mysql_probe_args(option_file: &std::path::Path) -> Vec<String> {
+    vec![
+        format!("--defaults-file={}", option_file.display()),
+        "--no-login-paths".to_string(),
+        "--protocol=TCP".to_string(),
+        "--connect-timeout=10".to_string(),
+        "--batch".to_string(),
+        "--raw".to_string(),
+        "--skip-column-names".to_string(),
+        "--binary-mode".to_string(),
+        "--skip-reconnect".to_string(),
+        format!("--execute={MYSQL_PROBE_QUERY}"),
+    ]
+}
+
+pub(super) async fn run_mysql_command<I, S>(
+    args: I,
+    option_file: Option<&PathBuf>,
+) -> Result<std::process::Output, DbOperationError>
+where
+    I: IntoIterator<Item = S>,
+    S: AsRef<OsStr>,
+{
+    let mut command = Command::new("mysql");
+    command
+        .args(args)
+        .stdin(Stdio::null())
+        .env_remove("MYSQL_PWD")
+        .env_remove("MYSQL_PASSWORD")
+        .kill_on_drop(true);
+    if option_file.is_some() {
+        command.stdout(Stdio::piped()).stderr(Stdio::piped());
+    }
+
+    match timeout(MYSQL_PROBE_TIMEOUT, command.output()).await {
+        Ok(Ok(output)) => Ok(output),
+        Ok(Err(error)) if error.kind() == io::ErrorKind::NotFound => {
+            Err(DbOperationError::CommandNotFound {
+                command: DatabaseCli::MySql,
+                details: error.to_string(),
+            })
+        }
+        Ok(Err(error)) => Err(DbOperationError::ConnectionFailed(error.to_string())),
+        Err(_) => Err(DbOperationError::Timeout(
+            "mysql probe exceeded the connection timeout".to_string(),
+        )),
+    }
+}
+
+pub(super) fn clean_stderr(stderr: &[u8]) -> String {
+    let text = String::from_utf8_lossy(stderr).trim().to_string();
+    if text.is_empty() {
+        "mysql probe failed".to_string()
+    } else {
+        text
+    }
+}
+
+pub(super) fn classify_mysql_probe_failure(stderr: String) -> DbOperationError {
+    if is_mysql_connect_timeout_message(&stderr) {
+        DbOperationError::Timeout(stderr)
+    } else {
+        DbOperationError::ConnectionFailed(stderr)
+    }
+}
+
+fn is_mysql_connect_timeout_message(value: &str) -> bool {
+    let lower = value.to_ascii_lowercase();
+    lower.contains("can't connect to mysql server")
+        && (lower.contains("(110)") || lower.contains("(10060)"))
+}
+
+#[cfg(test)]
+mod probe_tests {
+    use sabiql_app::model::connection::error::{ConnectionErrorInfo, ConnectionErrorKind};
+
+    use super::*;
+
+    #[test]
+    fn metadata_only_select_rejects_known_side_effects() {
+        for query in [
+            "SELECT value FROM items FOR UPDATE",
+            "SELECT GET_LOCK('sabiql', 0)",
+            "SELECT @value := 1",
+        ] {
+            assert!(
+                mysql_metadata_select_query(query, "__source", "__marker").is_err(),
+                "{query}"
+            );
+        }
+        assert!(mysql_metadata_select_query(
+            "WITH cte_rows AS (SELECT 1 AS first_alias) SELECT first_alias FROM cte_rows WHERE FALSE",
+            "__source",
+            "__marker"
+        )
+        .is_ok());
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT CONCAT('a', 'b') AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT CONCAT/**/('a', 'b') AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT sabiql_test.user_function() AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT sabiql_test/**/.user_function() AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT `user_function`/**/() AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT @session_value AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT INTERVAL(10, 1, 5) AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_err()
+        );
+        assert!(
+            mysql_metadata_select_query(
+            "WITH cte_rows(first_alias) AS (SELECT 1) SELECT first_alias FROM cte_rows WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_ok()
+        );
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT CASE (1) WHEN 1 THEN 'x' ELSE 'y' END AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_ok()
+        );
+        for query in [
+            "SELECT CAST(1 AS CHAR) AS value WHERE FALSE",
+            "SELECT CONVERT(1, CHAR) AS value WHERE FALSE",
+            "SELECT EXTRACT(YEAR FROM CURRENT_DATE) AS value WHERE FALSE",
+        ] {
+            assert!(
+                mysql_metadata_select_query(query, "__source", "__marker").is_err(),
+                "{query}"
+            );
+        }
+        assert!(
+            mysql_metadata_select_query(
+                "SELECT SLEEP(1) AS value WHERE FALSE",
+                "__source",
+                "__marker"
+            )
+            .is_ok()
+        );
+    }
+
+    #[test]
+    fn validates_supported_versions_and_sql_modes() {
+        assert!(is_oracle_mysql_cli_84_version("mysql  Ver 8.4.3 for macos"));
+        assert!(!is_oracle_mysql_cli_84_version(
+            "mysql  Ver 8.0.36 for macos"
+        ));
+        assert!(!is_oracle_mysql_cli_84_version(
+            "mysql  Ver 15.1 Distrib 10.11.8-MariaDB"
+        ));
+        assert!(!is_oracle_mysql_cli_84_version(
+            "mysql  Ver 8.4.3-3 for Linux (Percona Server)"
+        ));
+        assert!(is_oracle_mysql_server_84_version("8.4.3"));
+        assert!(!is_oracle_mysql_server_84_version("8.4.3-TiDB"));
+        assert!(!is_oracle_mysql_server_84_version("8.4.3-Percona"));
+        assert!(validate_sql_mode("STRICT_TRANS_TABLES").is_ok());
+        assert!(validate_sql_mode("STRICT_TRANS_TABLES,ANSI_QUOTES").is_err());
+    }
+
+    #[test]
+    fn uses_tcp_and_keeps_defaults_file_first() {
+        let args = mysql_probe_args(std::path::Path::new("/tmp/sabiql-mysql.cnf"));
+
+        assert_eq!(args[0], "--defaults-file=/tmp/sabiql-mysql.cnf");
+        assert_eq!(args[1], "--no-login-paths");
+        assert_eq!(args[2], "--protocol=TCP");
+        assert!(args.contains(&"--batch".to_string()));
+        assert!(args.contains(&"--raw".to_string()));
+        assert!(args.contains(&"--skip-column-names".to_string()));
+        assert!(args.contains(&"--binary-mode".to_string()));
+        assert!(args.contains(&"--skip-reconnect".to_string()));
+        assert!(
+            args.last()
+                .unwrap()
+                .starts_with("--execute=SELECT JSON_OBJECT")
+        );
+    }
+
+    #[test]
+    fn arguments_do_not_contain_credentials() {
+        let args = mysql_probe_args(std::path::Path::new("/tmp/sabiql-mysql.cnf"));
+
+        assert!(
+            args.iter()
+                .all(|argument| { !argument.contains("password") && !argument.contains("secret") })
+        );
+    }
+
+    #[test]
+    fn classifies_mysql_cli_timeout_errno_before_connection_refusal() {
+        assert!(matches!(
+            classify_mysql_probe_failure(
+                "ERROR 2003 (HY000): Can't connect to MySQL server on 'host:3306' (110)"
+                    .to_string()
+            ),
+            DbOperationError::Timeout(_)
+        ));
+        assert!(matches!(
+            classify_mysql_probe_failure(
+                "ERROR 2003 (HY000): Can't connect to MySQL server on 'host:3306' (111)"
+                    .to_string()
+            ),
+            DbOperationError::ConnectionFailed(_)
+        ));
+    }
+
+    #[test]
+    fn classifies_mysql_tls_probe_failure_for_connection_error() {
+        let error = classify_mysql_probe_failure(
+            "ERROR 2026 (HY000): SSL connection error: error:0A000086:SSL routines::certificate verify failed"
+                .to_string(),
+        );
+
+        assert_eq!(
+            ConnectionErrorInfo::from_db_operation_error(&error).kind,
+            ConnectionErrorKind::MySqlTlsHandshakeFailed
+        );
+    }
+}
