+++ src/infra/adapters/mysql/option_file.rs
@@ -231,7 +231,7 @@ mod tests {
         let (result, path) = {
             let option_file = MySqlOptionFile::create(&target()).unwrap();
             let path = option_file.path.clone();
-            let result = super::super::MysqlProcess::spawn_with_program(
+            let result = super::super::cli::MysqlProcess::spawn_with_program(
                 std::ffi::OsStr::new("__sabiql_missing_mysql_binary__"),
                 &path,
             );
